% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_checks.R
\name{check_is_tbl_df}
\alias{check_is_tbl_df}
\title{Checks if object is data frame or tibble}
\usage{
check_is_tbl_df(..., .abort_error = TRUE)
}
\arguments{
\item{...}{one or more objects to test}

\item{.abort_error}{Aborts with CLI message if any of the supplied objects are not dataframes or tibbles}
}
\value{
boolean values for each item(s) in .data
}
\description{
Ensures the passed object or objects are dataframes or tibbles and aborts with
an error if one of them is not.
}
