% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{codes_911}
\alias{codes_911}
\title{Internal regex and code patterns for nemsqar}
\format{
An object of class \code{character} of length 1.
}
\usage{
codes_911
}
\description{
This file contains internal constant objects used throughout the package.
These include regular expression patterns and lookup values that support
EMS clinical logic within nemsqar. Centralizing these values improves
maintainability and prevents duplication across functions.
}
\details{
All objects defined in this file are internal and are not exported.
}
\keyword{internal}
