% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_formals.R
\name{compare_formals}
\alias{compare_formals}
\title{Compare formals across functions}
\usage{
compare_formals(fun, others)
}
\arguments{
\item{fun}{A function object to analyze. Do not quote the function name.}

\item{others}{A list of function objects to compare against.}
}
\value{
A list with two elements:
\itemize{
\item \code{shared}: Character vector of argument names that appear in both the target
function and at least one comparison function.
\item unique: Character vector of argument names that appear only in the target
function.
}
}
\description{
Identify which arguments of a target function are unique and which are shared
with a set of comparison functions. Useful for determining whether roxygen2's
\verb{@inheritParams} is appropriate.
}
\examples{
compare_formals(
  airway_18_population,
  list(airway_05, airway_01)
)

}
\seealso{
Related functions.
}
\author{
Nicolas Foss, Ed.D., MS
}
