% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_normality.R
\name{compare_is_it_normal}
\alias{compare_is_it_normal}
\title{Compare Is It Normal}
\usage{
compare_is_it_normal(
  df,
  data_name = deparse(substitute(df)),
  x1,
  x2,
  include_plots = TRUE,
  normalize_x1 = FALSE,
  normalize_x2 = FALSE,
  filter_x1_1 = NULL,
  filter_x1_2 = NULL,
  operator_x1 = NULL,
  filter_x2_1 = NULL,
  filter_x2_2 = NULL,
  operator_x2 = NULL,
  print_out = TRUE,
  plot_theme = theme_cleaner,
  ...
)
}
\arguments{
\item{df}{Data frame or tibble containing the variables to compare.}

\item{data_name}{Optional name of the dataset (default is the df object name).}

\item{x1}{The first numeric variable to assess for normality.}

\item{x2}{The second numeric variable to assess for normality.}

\item{include_plots}{Logical flag to include diagnostic plots (TRUE by default).}

\item{normalize_x1}{Logical flag to normalize x1 (min-max normalization).}

\item{normalize_x2}{Logical flag to normalize x2 (min-max normalization).}

\item{filter_x1_1}{First filtering condition for x1 as a RHS logical argument as a string (e.g. " < 40").}

\item{filter_x1_2}{Second filtering condition for x1 as a RHS logical argument as a string (e.g. " < 50").}

\item{operator_x1}{Logical operator for combining filter conditions for x1 as a string (e.g. " & ").}

\item{filter_x2_1}{First filtering condition for x2 as a RHS logical argument as a string (e.g. " < 40").}

\item{filter_x2_2}{Second filtering condition for x2 as a RHS logical argument as a string (e.g. " < 50").}

\item{operator_x2}{Logical operator for combining filter conditions for x2 as a string (e.g. " & ").}

\item{print_out}{Logical flag to print results to the console.}

\item{plot_theme}{Custom ggplot2 theme for the diagnostic plots (theme_cleaner by default).}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
Combined plots or tibble
}
\description{
Comparative function for two variables from the same df
}
\details{
The \code{compare_is_it_normal} function is designed to assess the normality of
two numeric variables from the same dataset. It compares these variables using
the Shapiro-Wilk test, provides descriptive statistics, and generates visual
diagnostics such as Q-Q plots and histograms. This function is a valuable tool
for exploratory data analysis when investigating whether two distributions
follow a normal distribution.
}
\section{details}{

The \code{compare_is_it_normal} function provides a comprehensive analysis of two
numeric variables within the same dataset. It supports the following steps:
\itemize{
\item Checks the class of input variables and applies optional filtering.
\item Normalizes the variables if required.
\item Performs the Shapiro-Wilk test for normality.
\item Computes descriptive statistics including mean, standard deviation, and
percent missing data.
\itemize{
\item Generates diagnostic plots such as Q-Q plots and histograms if
\code{include_plots = TRUE}.
}
}
}

\section{notes}{

\itemize{
\item When the sample size exceeds 5,000, the Shapiro-Wilk test is conducted on a
random sample of 5,000 records to ensure computational efficiency.
\item Diagnostic plots (Q-Q plots, histograms) are only generated if
include_plots = TRUE.
\item Filters can be applied to both variables to limit the range of the data
being analyzed, but must be passed as RHS logical arguments as a string
(see \emph{Arguments} above).  If more than one filter is applied, it must be
joined by an \code{operator}, which is a string logical operator such as " & "
or " | ".
\item This plot uses \code{theme_cleaner} which is a custom ggplot2 theme created for
\code{traumaR}, and can be accessed via this Github repo.
}
}

