% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify.R
\name{format_p_value}
\alias{format_p_value}
\title{Format p Value}
\usage{
format_p_value(p_value)
}
\arguments{
\item{p_value}{numeric p-value to be formatted}
}
\value{
string containing p-value significance and p-value.
}
\description{
Prints a prettified p value that groups on statistical significance
}
\section{Credit}{


This function was developed by (Nicolas Foss, Ed.D., MS)\link{nicolas.foss@hhs.iowa.gov} at the Bureau of Emergency Medical and Trauma
Services, Division of Public Health, Iowa HHS.
}

\examples{
format_p_value(0.12)
# "p > 0.05 (p = 0.12)"
format_p_value(0.0123)
# "p < 0.05 (p = 0.0123)"
}
