% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Not In operator}
\usage{
x \%not_in\% y
}
\arguments{
\item{x}{A vector whose elements will be checked.}

\item{y}{A vector against which to check for the presence of elements from x.}
}
\value{
Returns a logical vector indicating \code{TRUE} for elements of \code{x} that
are not in \code{y}, and \code{FALSE} for those that are.
}
\description{
The \verb{\%not_in\%} function is an inverse operator that checks if elements of one
vector are not present in another vector. It returns a logical vector
indicating the absence of each element.
}
\details{
The operator \verb{\%not_in\%} is a user-friendly alternative to the standard negation
of the \%in\% operator.
}
\section{Credit}{


This function was developed by (Nicolas Foss, Ed.D., MS)\link{nicolas.foss@hhs.iowa.gov} at the Bureau of Emergency Medical and Trauma
Services, Division of Public Health, Iowa HHS.
}

\examples{
x <- c(1, 2, 3, 4)
y <- c(2, 4, 6)

x \%not_in\% y
# TRUE FALSE TRUE FALSE

}
