% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_normality.R
\name{is_it_normal}
\alias{is_it_normal}
\title{Is It Normal}
\usage{
is_it_normal(
  df,
  data_name = deparse(substitute(df)),
  x,
  include_plots = TRUE,
  filter1 = NULL,
  filter2 = NULL,
  operator = NULL,
  scale = FALSE,
  print_out = TRUE,
  plot_theme = theme_cleaner,
  ...
)
}
\arguments{
\item{df}{A dataframe or tibble containing the variable to be analyzed.}

\item{data_name}{Optional name of dataset.}

\item{x}{(Required) The unquoted name of the numeric variable to test for normality.}

\item{include_plots}{(Logical) Whether to include visualizations such as Q-Q plot, histogram, density plot, and boxplot. Default is TRUE.}

\item{filter1}{(Optional) A condition to filter the dataframe based on the variable. Used when you want to filter data before the analysis.  Must be a RHS logical argument as a string (e.g. " < 40").}

\item{filter2}{(Optional) A second condition for filtering.  Must be a RHS logical argument as a string (e.g. " > 50").}

\item{operator}{(Optional) A logical operator (e.g., &, |) for more complex filtering conditions as a string to join \code{filter1} and \code{filter2}.}

\item{scale}{(Logical) Whether to apply Min-Max normalization to the variable before the analysis. Default is FALSE.}

\item{print_out}{(Logical) Whether to print the output tibble or return it invisibly. Default is TRUE.}

\item{plot_theme}{(Optional) A ggplot2 theme function name to apply custom styling to the plots. Default is theme_cleaner.}

\item{...}{Additional arguments passed to ggplot2 themes or other customization.}
}
\value{
A tibble containing mean, SD, min/max/quartiles, number of missing values, Shapiro-Wilk p-value, normality diagnosis, and optional plots.
}
\description{
Assesses whether a continuous numeric variable in a dataset follows a normal distribution.
}
\details{
The \code{is_it_normal} function is designed to assess whether a continuous numeric
variable in a dataset follows a normal distribution. This function provides
key descriptive statistics, performs a Shapiro-Wilk test for normality, and
offers a range of visualizations such as Q-Q plots, histograms, kernel density
plots, and boxplots to help you evaluate the distribution of the variable.
}
\section{details}{

The is_it_normal function provides a comprehensive way to check whether a continuous variable in a dataset is normally distributed. It performs several key tasks:
\itemize{
\item Filters the dataframe based on provided conditions.
\item Computes descriptive statistics such as mean, standard deviation, quartiles, and missing data information.
\item Conducts a Shapiro-Wilk test of normality.
\item Optionally applies Min-Max scaling to the data.
\item Generates several useful visualizations including:
\itemize{
\item Q-Q Plot: to visualize the quantiles against a theoretical normal distribution.
\item Histogram: to observe the shape of the distribution.
\item Kernel Density Plot: to estimate the probability density function.
\item Boxplot: to summarize the distribution through median and quartiles.
}
}
}

\examples{
data(mtcars)
# Check if 'mpg' variable is normallly distributed.
is_it_normal(mtcars, mpg)

# Using a filter and custom theme
is_it_normal(mtcars, mpg, filter1 = " > 15", filter2 = " < 100", operator = " & ", plot_theme = theme_minimal)
}
