% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemsqar-package.R
\docType{package}
\name{nemsqar-package}
\alias{nemsqar}
\alias{nemsqar-package}
\title{nemsqar: National Emergency Medical Service Quality Alliance Measure Calculations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Designed to automate the calculation of Emergency Medical Service (EMS) quality metrics, 'nemsqar' implements measures defined by the National EMS Quality Alliance (NEMSQA). By providing reliable, evidence-based quality assessments, the package supports EMS agencies, healthcare providers, and researchers in evaluating and improving patient outcomes. Users can find details on all approved NEMSQA measures at \url{https://www.nemsqa.org/measures}. Full technical specifications, including documentation and pseudocode used to develop 'nemsqar', are available on the NEMSQA website after creating a user profile at \url{https://www.nemsqa.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bemts-hhs/nemsqar}
  \item \url{https://bemts-hhs.github.io/nemsqar/}
  \item Report bugs at \url{https://github.com/bemts-hhs/nemsqar/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Foss \email{nicolas.foss@hhs.iowa.gov}

Authors:
\itemize{
  \item Nicolas Foss \email{nicolas.foss@hhs.iowa.gov}
  \item Samuel Kordik \email{samuel.kordik@utsouthwestern.edu} (\href{https://orcid.org/0000-0003-4230-1154}{ORCID})
}

Other contributors:
\itemize{
  \item Alyssa Green \email{agreen@nemsqa.org} [contributor]
  \item Iowa Department of Health and Human Services [copyright holder]
}

}
\keyword{internal}
