% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemsqar_airway_01_df.R
\docType{data}
\name{nemsqar_airway_01_df}
\alias{nemsqar_airway_01_df}
\title{Synthetic Test Data for Airway-18 Measure in National EMS Information System}
\format{
A data frame with 1,000 rows and 16 variables:
\describe{
\item{Incident Patient Care Report Number - PCR (eRecord.01)}{Unique identifier for the incident report (character).}
\item{Incident Date}{Date of the incident (Date).}
\item{Patient Age (ePatient.15)}{Age of the patient (numeric).}
\item{Patient Age Units (ePatient.16)}{Unit of measurement for patient age (character).}
\item{Patient Date Of Birth (ePatient.17)}{Patient's date of birth (Date).}
\item{Patient Gender (ePatient.13)}{Gender of the patient (character).}
\item{Response Type Of Service Requested With Code (eResponse.05)}{Type of service requested for the EMS response (character).}
\item{Cardiac Arrest During EMS Event With Code (eArrest.01)}{Indicates whether a cardiac arrest occurred during the EMS event (character).}
\item{Vitals Signs Taken Date Time (eVitals.01)}{Timestamp when vital signs were recorded (datetime).}
\item{Vitals Systolic Blood Pressure SBP (eVitals.06)}{Recorded systolic blood pressure (numeric).}
\item{Vitals Pulse Oximetry (eVitals.12)}{Oxygen saturation level (numeric).}
\item{Procedure Performed Date Time (eProcedures.01)}{Timestamp when the procedure was performed (datetime).}
\item{Procedure Performed Prior To EMS Care (eProcedures.02)}{Indicates if the procedure was performed before EMS arrival (character).}
\item{Procedure Performed Description And Code (eProcedures.03)}{Description and standardized code of the performed procedure (character).}
\item{Procedure Number Of Attempts (eProcedures.05)}{Number of attempts made for the procedure (numeric).}
\item{Procedure Successful (eProcedures.06)}{Indicates whether the procedure was successful (character).}
}
}
\usage{
data(nemsqar_airway_01_df)
}
\description{
This dataset provides completely synthetic test data for evaluating the \code{airway_01}
and \code{airway_01_population} functions in the \code{nemsqar} package. It contains fields relevant
to airway management and procedural success, allowing users to test the functionality of
these measure functions.
}
\details{
Users are encouraged to test these functions with this dataset, but results
should not be interpreted as meaningful. Some outputs may be nonsensical,
which is expected since this data is only intended to demonstrate the
expected structure of input data.
}
\examples{
data(nemsqar_airway_01_df)
head(nemsqar_airway_01_df)

}
\keyword{datasets}
