% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemsqar_airway_18_df.R
\docType{data}
\name{nemsqar_airway_18_df}
\alias{nemsqar_airway_18_df}
\title{Synthetic Test Data for Airway-18 Measure}
\format{
A data frame with 1,000 rows and 16 variables:
\describe{
\item{Incident Patient Care Report Number - PCR (eRecord.01)}{Unique identifier for the incident report (character).}
\item{Incident Date}{Date of the incident (Date).}
\item{Patient Age (ePatient.15)}{Patient's age (numeric).}
\item{Patient Age Units (ePatient.16)}{Units for patient age (character).}
\item{Patient Date Of Birth (ePatient.17)}{Patient's birth date (Date).}
\item{Patient Gender (ePatient.13)}{Gender of the patient (character).}
\item{Procedure Performed Description And Code (eProcedures.03)}{Description and SNOMED code of procedure performed (character).}
\item{Procedure Number Of Attempts (eProcedures.05)}{Number of attempts for the procedure (numeric).}
\item{Procedure Successful (eProcedures.06)}{Whether the procedure was successful (character).}
\item{Response Type Of Service Requested With Code (eResponse.05)}{Type of EMS response requested (character).}
\item{Procedure Performed Prior To EMS Care (eProcedures.02)}{Indicates if the procedure was performed before EMS arrival (character).}
\item{Procedure Performed Date Time (eProcedures.01)}{Timestamp of procedure performance (character).}
\item{Airway Device Placement Confirmed Method List (eAirway.04)}{Methods used to confirm airway device placement (character).}
\item{Airway Device Placement Confirmation Date Time (eAirway.02)}{Timestamp of airway device placement confirmation (datetime).}
\item{Vitals Signs Taken Date Time (eVitals.01)}{Timestamp of vitals assessment (character).}
\item{Vitals Carbon Dioxide CO2 (eVitals.16)}{Carbon dioxide measurement in vitals (numeric).}
}
}
\usage{
data(nemsqar_airway_18_df)
}
\description{
This dataset provides completely synthetic test data for evaluating the
\code{airway_18} and \code{airway_18_population} functions in the \code{nemsqar} package.
Users are encouraged to test these functions with this dataset, but results
should not be interpreted as meaningful. Some outputs may be nonsensical,
which is expected since this data is only intended to demonstrate the
expected structure of input data.
}
\examples{
data(nemsqar_airway_18_df)
head(nemsqar_airway_18_df)

}
\keyword{datasets}
