% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orRify}
\alias{orRify}
\title{orRify}
\usage{
orRify(string)
}
\arguments{
\item{string}{A character vector containing the elements to be concatenated with the "or" operator.}
}
\value{
Returns a single concatenated string, with each element separated by a \code{|} operator.
}
\description{
The \code{orRify} function concatenates a vector of strings with the "or" operator
(\code{|}) between each element, producing a single output string. This function
is particularly useful for constructing search patterns or queries that r
equire an "or" condition between elements.
}
\details{
The function assumes the input vector has at least one element; if the vector
is empty, it will return an empty string.
}
\examples{
terms <- c("apple", "banana", "cherry")
orRified_terms <- orRify(terms)
print(orRified_terms)
# "apple|banana|cherry"

}
