% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify.R
\name{pretty_number}
\alias{pretty_number}
\title{Pretty Number}
\usage{
pretty_number(x, n_decimal = 2, prefix = NULL, round = T)
}
\arguments{
\item{x}{A numeric vector to be formatted.}

\item{n_decimal}{The number of decimal places to round to (default is 2).}

\item{prefix}{An optional character string to prepend to the formatted number.}

\item{round}{A logical value indicating whether to round the numbers (default is \code{TRUE}).}
}
\value{
Returns a character vector of formatted numbers.
}
\description{
Formats large numeric values into a more human-readable string.
}
\details{
The \code{pretty_number} function formats large numeric values into a more readable
string representation with optional decimal rounding and prefix support. It
converts numbers into a compact format using standard suffixes (e.g., "k" for
thousand, "m" for million).
}
\section{Credit}{


This function was developed by (Nicolas Foss, Ed.D., MS)\link{nicolas.foss@hhs.iowa.gov} at the Bureau of Emergency Medical and Trauma
Services, Division of Public Health, Iowa HHS.
}

\examples{
pretty_number(1500)
# "1.5k"

pretty_number(2500000, prefix = "$")
# "$2.5m"

}
