% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify.R
\name{pretty_percent}
\alias{pretty_percent}
\title{Pretty Percent}
\usage{
pretty_percent(variable, n_decimal = 0.1)
}
\arguments{
\item{variable}{A numeric vector representing the values to be formatted as percentages.}

\item{n_decimal}{The precision (number of decimal places, scale from 0 to 1) to which the percentage is formatted (default is 0.1).}
}
\value{
string representation of percentage
}
\description{
Function to clean up percent formatted numbers
}
\details{
The function utilizes the percent function from the scales package to format
the input as a percentage.It removes trailing zeros after the decimal point
and simplifies formatting by eliminating unnecessary characters.
}
\examples{
numerator <- c(7.5, 20.0, 0.3333, 100.0, 15.75, 0.005, 150.0)
denominator <- c(100.0, 100.0, 1.0, 100.0, 100.0, 0.1, 300.0)

percent_values <- numerator / denominator  # This will create values like 0.075, 0.20, etc.
formatted_percent <- pretty_percent(percent_values, n_decimal = 0.01)
print(formatted_percent)
#Output: "7.5\%", "20\%", "33.33\%", "100\%", "15.75\%", "5\%", "50\%"

}
\author{
Nicolas Foss, Ed.D., MS
}
