% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airway_01.R
\name{airway_01}
\alias{airway_01}
\title{Airway-01 Calculation}
\usage{
airway_01(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  arrest_table = NULL,
  procedures_table = NULL,
  vitals_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  earrest_01_col,
  eresponse_05_col,
  evitals_01_col,
  evitals_06_col,
  evitals_12_col,
  eprocedures_01_col,
  eprocedures_02_col,
  eprocedures_03_col,
  eprocedures_05_col,
  eprocedures_06_col,
  ...
)
}
\arguments{
\item{df}{A dataframe or tibble contianing EMS data where each row represents
an observation and columns represent features.}

\item{patient_scene_table}{A data.frame or tibble containing at least
epatient, escene, and earrest.01 fields as a fact table.}

\item{response_table}{A data.frame or tibble containing at least the
eresponse fields needed for this measure's calculations.}

\item{arrest_table}{A data.frame or tibble containing at least the earrest
fields needed for this measure's calculations.}

\item{procedures_table}{A dataframe or tibble containing at least the
eProcedures fields needed.}

\item{vitals_table}{A dataframe or tibble containing at least the eVitals
fields needed.}

\item{erecord_01_col}{The column representing the EMS record unique
identifier.}

\item{incident_date_col}{Column that contains the incident date. This
defaults to \code{NULL} as it is optional in case not available due to PII
restrictions.}

\item{patient_DOB_col}{Column that contains the patient's date of birth. This
defaults to \code{NULL} as it is optional in case not available due to PII
restrictions.}

\item{epatient_15_col}{Column representing the patient's numeric age agnostic
of unit.}

\item{epatient_16_col}{Column representing the patient's age unit ("Years",
"Months", "Days", "Hours", or "Minutes").}

\item{earrest_01_col}{Column representing whether or not the patient is in
arrest.}

\item{eresponse_05_col}{Column that contains eResponse.05.}

\item{evitals_01_col}{Date-time or POSIXct column containing vital signs
date/time}

\item{evitals_06_col}{Numeric column containing systolic blood pressure
values}

\item{evitals_12_col}{Numeric column containing pulse oximetry values.}

\item{eprocedures_01_col}{Date-time or POSIXct column for procedures}

\item{eprocedures_02_col}{Column name for whether or not the procedure was
performed prior to EMS care being provided.}

\item{eprocedures_03_col}{Column containing procedure codes with or without
procedure names.}

\item{eprocedures_05_col}{Column containing a count for how many times
procedure was attempted.}

\item{eprocedures_06_col}{Column indicating whether or not procedure was
successful.}

\item{...}{optional additional arguments to pass onto \code{dplyr::summarize}.}
}
\value{
A data.frame summarizing results for three population groups (All,
Adults, and Peds) with the following columns:
\code{pop}: Population type (All, Adults, or Peds).
\code{numerator}: Count of incidents where beta-agonist medications were
administered.
\code{denominator}: Total count of incidents.
\code{prop}: Proportion of incidents involving beta-agonist medications.
\code{prop_label}: Proportion formatted as a percentage with a specified number of
decimal places.
}
\description{
Calculates the NEMSQA Airway-01 measure.

Calculates the proportion of times when the first endotracheal intubation
attempt is successful with no peri-intubation hypoxia or hypotension.
}
\examples{

\dontrun{

# If you are sourcing your data from a SQL database connection
# or if you have your data in several different tables,
# you can pass table inputs versus a single data.frame or tibble

# Get the applicable tables from `nemsqar`
data("nemsqar_arrest_table")
data("nemsqar_patient_scene_table")
data("nemsqar_response_table")
data("nemsqar_vitals_table")
data("nemsqar_procedures_table")

# Run the function

airway_01(df = NULL,
         patient_scene_table = nemsqar_patient_scene_table,
         procedures_table = nemsqar_procedures_table,
         vitals_table = nemsqar_vitals_table,
         arrest_table = nemsqar_arrest_table,
         response_table = nemsqar_response_table,
         erecord_01_col = `Incident Patient Care Report Number - PCR (eRecord.01)`,
         incident_date_col = `Incident Date`,
         patient_DOB_col = `Patient Date Of Birth (ePatient.17)`,
         epatient_15_col = `Patient Age (ePatient.15)`,
         epatient_16_col = `Patient Age Units (ePatient.16)`,
         eresponse_05_col = `Response Type Of Service Requested With Code (eResponse.05)`,
         eprocedures_01_col = `Procedure Performed Date Time (eProcedures.01)`,
         eprocedures_02_col = `Procedure Performed Prior To EMS Care (eProcedures.02)`,
         eprocedures_03_col = `Procedure Performed Description And Code (eProcedures.03)`,
         eprocedures_05_col = `Procedure Number Of Attempts (eProcedures.05)`,
         eprocedures_06_col = `Procedure Successful (eProcedures.06)`,
         earrest_01_col = `Cardiac Arrest During EMS Event With Code (eArrest.01)`,
         evitals_01_col = `Vitals Signs Taken Date Time (eVitals.01)`,
         evitals_06_col = `Vitals Systolic Blood Pressure SBP (eVitals.06)`,
         evitals_12_col = `Vitals Pulse Oximetry (eVitals.12)`
         )
}

}
\author{
Samuel Kordik, Nicolas Foss Ed.D., MS
}
