% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airway_18.R
\name{airway_18}
\alias{airway_18}
\title{Airway-18}
\usage{
airway_18(
  df,
  erecord_01_col,
  incident_date_col,
  patient_DOB_col,
  epatient_15_col,
  epatient_16_col,
  eprocedures_03_col,
  eresponse_05_col,
  eprocedures_06_col,
  eprocedures_01_col,
  eprocedures_02_col,
  eairway_04_col,
  eairway_02_col,
  evitals_01_col,
  evitals_16_col,
  ...
)
}
\arguments{
\item{df}{A data frame or tibble containing the dataset to be processed.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name containing the unique patient record identifier.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name containing the incident date.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name containing the patient's date of birth.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for patient information (exact purpose unclear).}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for patient information (exact purpose unclear).}

\item{eprocedures_03_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for procedure codes.}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for emergency response codes.}

\item{eprocedures_06_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for procedure success codes.}

\item{eprocedures_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for procedure times or other related data.}

\item{eprocedures_02_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for additional procedure data.}

\item{eairway_04_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for airway procedure data.}

\item{eairway_02_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for airway procedure data (datetime).}

\item{evitals_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for vital signs data (datetime).}

\item{evitals_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for additional vital signs data.}

\item{...}{Additional arguments passed to other functions if needed.}
}
\value{
A tibble summarizing results for Adults and Peds with the following columns:
\code{pop}: Population type (Adults, Peds).
\code{numerator}: Count of incidents where waveform capnography is used for tube placement confirmation on
the last successful invasive airway procedure.
\code{denominator}: Total count of incidents.
\code{prop}: Proportion of incidents where waveform capnography is used for tube placement confirmation on
the last successful invasive airway procedure.
\code{prop_label}: Proportion formatted as a percentage with a specified number of
decimal places.
}
\description{
This function processes and analyzes the dataset to calculate the "Airway-18" NEMSQA metric.
It includes cleaning and transforming several columns related to patient data, airway procedures,
and vital signs, and it returns a cleaned dataset with the relevant calculations.
The final calculation is an assessment of the successful last invasive airway procedures
performed during an EMS response originating from a 911 request in which waveform capnography
is used for tube placement confirmation.
}
\note{
This function filters and processes EMS data to:
\itemize{
\item Identify the last successful airway procedure.
\item Filter for 911 response codes and relevant vital signs (i.e. ETCO2).
\item Aggregate results by patient encounter and calculate stroke scale outcomes.
\item Return a summary of stroke cases by unique patient identifier, including stroke scale measurements.
}
}
\section{Data Assumptions}{


#' This function assumes that:
Data are already loaded. The data needs to be a data.frame or tibble.

Age in years will be calculated using the patient date of birth and incident
date. These fields must have valid Date or POSIXct data types.

When values are missing, they are coded as NA, not the "not known"/"not
recorded" values common to ImageTrend or the NEMSIS codes that correspond to
"not values".

For the eprocedure, eairway, and evitals fields, all responses entered can be
included in \code{df}.  This will result in Cartesian product, but the function
will performantly handle that problem.
}

\section{Practical Tips}{


Ensure data are pre-processed, with missing values coded as \code{NA}, before passing
into the function.
Prepare necessary joins (e.g., for vitals) in advance; this function does not perform joins.
}

\author{
Nicolas Foss, Ed.D., MS
}
