% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airway_18_population.R
\name{airway_18_population}
\alias{airway_18_population}
\title{Airway-18 Populations}
\usage{
airway_18_population(
  df = NULL,
  patient_scene_table = NULL,
  procedures_table = NULL,
  vitals_table = NULL,
  airway_table = NULL,
  response_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  eprocedures_01_col,
  eprocedures_02_col,
  eprocedures_03_col,
  eprocedures_05_col,
  eprocedures_06_col,
  eairway_02_col = NULL,
  eairway_04_col = NULL,
  evitals_01_col,
  evitals_16_col
)
}
\arguments{
\item{df}{A data frame or tibble containing the dataset to be processed.
Default is \code{NULL}.}

\item{patient_scene_table}{A data frame or tibble containing only ePatient
and eScene fields as a fact table. Default is \code{NULL}.}

\item{procedures_table}{A data frame or tibble containing only the
eProcedures fields needed for this measure's calculations. Default is
\code{NULL}.}

\item{vitals_table}{A data frame or tibble containing only the eVitals fields
needed for this measure's calculations. Default is \code{NULL}.}

\item{airway_table}{A data frame or tibble containing only the eAirway fields
needed for this measure's calculations. Default is \code{NULL}.}

\item{response_table}{A data frame or tibble containing only the eResponse
fields needed for this measure's calculations. Default is \code{NULL}.}

\item{erecord_01_col}{Column name containing the unique patient record
identifier.}

\item{incident_date_col}{Column that contains the incident date. This
defaults to \code{NULL} as it is optional in case not available due to PII
restrictions.}

\item{patient_DOB_col}{Column that contains the patient's date of birth. This
defaults to \code{NULL} as it is optional in case not available due to PII
restrictions.}

\item{epatient_15_col}{Column name for patient information (exact purpose
unclear).}

\item{epatient_16_col}{Column name for patient information (exact purpose
unclear).}

\item{eresponse_05_col}{Column name for emergency response codes.}

\item{eprocedures_01_col}{Column name for procedure times or other related
data.}

\item{eprocedures_02_col}{Column name for whether or not the procedure was
performed prior to EMS care being provided.}

\item{eprocedures_03_col}{Column name for procedure codes.}

\item{eprocedures_05_col}{Column name for number of procedure attempts.}

\item{eprocedures_06_col}{Column name for procedure success codes.}

\item{eairway_02_col}{Column name for airway procedure data (datetime).
Default is \code{NULL}.}

\item{eairway_04_col}{Column name for airway procedure data. Default is
\code{NULL}.}

\item{evitals_01_col}{Column name for vital signs data (datetime).}

\item{evitals_16_col}{Column name for additional vital signs data.}
}
\value{
A list that contains the following:
\itemize{
\item a tibble with counts for each filtering step,
\item a tibble for each population of interest
\item a tibble for the initial population
\item a tibble for the total dataset with computations
}
}
\description{
This function processes and analyzes the dataset to generate the populations
of interest needed to perform calculations to obtain performance data.
}
\examples{

\dontrun{

# If you are sourcing your data from a SQL database connection
# or if you have your data in several different tables,
# you can pass table inputs versus a single data.frame or tibble

# Get the applicable tables from `nemsqar`
data("nemsqar_airway_table")
data("nemsqar_patient_scene_table")
data("nemsqar_response_table")
data("nemsqar_vitals_table")
data("nemsqar_procedures_table")

# Run the function

airway_18_population(df = NULL,
         patient_scene_table = nemsqar_patient_scene_table,
         procedures_table = nemsqar_procedures_table,
         vitals_table = nemsqar_vitals_table,
         airway_table = nemsqar_airway_table,
         response_table = nemsqar_response_table,
         erecord_01_col = `Incident Patient Care Report Number - PCR (eRecord.01)`,
         incident_date_col = `Incident Date`,
         patient_DOB_col = `Patient Date Of Birth (ePatient.17)`,
         epatient_15_col = `Patient Age (ePatient.15)`,
         epatient_16_col = `Patient Age Units (ePatient.16)`,
         eresponse_05_col = `Response Type Of Service Requested With Code (eResponse.05)`,
         eprocedures_01_col = `Procedure Performed Date Time (eProcedures.01)`,
         eprocedures_02_col = `Procedure Performed Prior To EMS Care (eProcedures.02)`,
         eprocedures_03_col = `Procedure Performed Description And Code (eProcedures.03)`,
         eprocedures_05_col = `Procedure Number Of Attempts (eProcedures.05)`,
         eprocedures_06_col = `Procedure Successful (eProcedures.06)`,
         eairway_02_col = `Airway Device Placement Confirmation Date Time (eAirway.02)`,
         eairway_04_col = `Airway Device Placement Confirmed Method List (eAirway.04)`,
         evitals_01_col = `Vitals Signs Taken Date Time (eVitals.01)`,
         evitals_16_col = `Vitals Carbon Dioxide CO2 (eVitals.16)`
         )
}

}
\author{
Nicolas Foss, Ed.D., MS, Samuel Kordik, BBA, BS
}
