% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asthma_01.R
\name{asthma_01}
\alias{asthma_01}
\title{Asthma-01}
\usage{
asthma_01(
  df,
  incident_date_col,
  patient_DOB_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  emedications_03_col,
  ...
)
}
\arguments{
\item{df}{A data.frame or tibble containing EMS data where each row represents
an observation, and columns represent features.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains the incident date.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains the patient's date of birth.}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains eResponse.05.}

\item{esituation_11_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains eSituation.11.}

\item{esituation_12_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains all eSituation.12 values as a single comma-separated list.}

\item{emedications_03_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains all eMedications.03 values as a single comma-separated list.}

\item{...}{optional additional arguments to pass onto \code{dplyr::summarize}.}
}
\value{
A data.frame summarizing results for three population groups (All,
Adults, and Peds) with the following columns:
\code{pop}: Population type (All, Adults, or Peds).
\code{numerator}: Count of incidents where beta-agonist medications were administered.
\code{denominator}: Total count of incidents.
\code{prop}: Proportion of incidents involving beta-agonist medications.
\code{prop_label}: Proportion formatted as a percentage with a specified number of
decimal places.
}
\description{
Calculates the NEMSQA Asthma-01 measure.
}
\details{
Calculates key statistics related to asthma-related incidents in an EMS dataset,
specifically focusing on cases where 911 was called for respiratory distress,
and certain medications were administered. This function segments the data by
age into adult and pediatric populations, computing the proportion of cases that
received beta-agonist treatment.
}
\section{Data Assumptions}{


This function assumes that:
Data are already loaded. The data needs to be a dataframe where each row is
one observation (patient) and each is a feature (field) or distinct datasets
that can be references as unique columns.

Age in years will be calculated using the patient date of birth and incident
date. These fields must have valid Date or POSIXct data types.

When values are missing, they are coded as NA, not the "not known"/"not
recorded" values common to ImageTrend or the NEMSIS codes that correspond to
"not values".

The primary and secondary impression fields (eSituation.11 and eSituation.12)
have the ICD-10 codes present in them. These fields may optionally contain
text for reference. Similarly, this function assumes that the eResponse.05
column has NEMSIS codes present, but text can also be included for reference.

The eMedications.03 field contains all medications administered and contains
a text description of the medication using the generic name. The RxNORM code
may also be included for reference, but will not be checked.

The secondary impressions field (eSituation.12) is best prepared as a
comma-separated list of all values in a single string.
}

\section{Practical Tips}{


The first argument is the dataframe prepared as above. No joining is done.
Any joins to get vitals, etc. will need to be done outside of this function.
Grouping can be done before the function to get the calculations by region or
other groupings.
}

\section{Features}{

\itemize{
\item Filters for asthma-related incidents (ICD-10 codes starting with 'J45' and
'J98.01').
\item Distinguishes between adults (age ≥ 18) and pediatric patients (age 2–17).
Calculates age in years based on incident date and patient date of birth.
Formats proportions as percentages with customizable decimal precision.
}
}

\section{Value}{

A summarized data.frame with counts and proportions of beta-agonist treatment
among 911 respiratory distress calls, segmented by population group.
}

