% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asthma_01_population.R
\name{asthma_01_population}
\alias{asthma_01_population}
\title{Asthma-01 Populations}
\usage{
asthma_01_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  medications_table = NULL,
  erecord_01_col,
  incident_date_col,
  patient_DOB_col,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  emedications_03_col
)
}
\arguments{
\item{df}{A data.frame or tibble containing EMS data where each row represents
an observation, and columns represent features.}

\item{patient_scene_table}{A data.frame or tibble containing only epatient and escene fields as a fact table.}

\item{response_table}{A data.frame or tibble containing only the eresponse fields needed for this measure's calculations.}

\item{situation_table}{A data.frame or tibble containing only the esituation fields needed for this measure's calculations.}

\item{medications_table}{A data.frame or tibble containing only the emedications fields needed for this measure's calculations.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column representing the EMS record unique identifier.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains the incident date.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains the patient's date of birth.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column representing the patient's numeric age agnostic of unit.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column representing the patient's age unit ("Years", "Months", "Days", "Hours", or "Minute").}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains eResponse.05.}

\item{esituation_11_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains eSituation.11.}

\item{esituation_12_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains all eSituation.12 values as a single comma-separated list.}

\item{emedications_03_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains all eMedications.03 values as a single comma-separated list.}
}
\value{
#' A list that contains the following:
\itemize{
\item a tibble with counts for each filtering step,
\item a tibble for each population of interest
\item a tibble for the initial population
}
}
\description{
Filters data down to the target populations for Asthma-01, and categorizes
records to identify needed information for the calculations.
}
\details{
Identifies key categories related to asthma-related incidents in an EMS dataset,
specifically focusing on cases where 911 was called for respiratory distress,
and certain medications were administered. This function segments the data by
age into adult and pediatric populations, computing the proportion of cases that
received beta-agonist treatment.
}
\section{Data Assumptions}{


This function assumes that:
Data are already loaded. The data needs to be a data.frame or tibble where each row is
one observation (patient) and each is a feature (field) or distinct datasets
that can be references as unique columns.

Age in years will be calculated using the patient date of birth and incident
date. These fields must have valid Date or POSIXct data types.

When values are missing, they are coded as NA, not the "not known"/"not
recorded" values common to ImageTrend or the NEMSIS codes that correspond to
"not values".

The primary and secondary impression fields (eSituation.11 and eSituation.12)
have the ICD-10 codes present in them. These fields may optionally contain
text for reference. Similarly, this function assumes that the eResponse.05
column has NEMSIS codes present, but text can also be included for reference.

The eMedications.03 field contains all medications administered and contains
a text description of the medication using the generic name. The RxNORM code
may also be included for reference, but will not be checked.

The secondary impressions field (eSituation.12) is best prepared as a
comma-separated list of all values in a single string.
}

\section{Practical Tips}{


The first argument is the dataframe or tables with data elements prepared as above.
No joining is done. Any joins to get vitals, etc. will need to be done outside of this function.
}

\section{Features}{

\itemize{
\item Filters for asthma-related incidents (ICD-10 codes starting with 'J45' and
'J98.01').
\item Distinguishes between adults (age ≥ 18) and pediatric patients (age 2–17).
Calculates age in years based on incident date and patient date of birth.
}
}

\author{
Nicolas Foss, Ed.D., MS
}
