% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoglycemia_01.R
\name{hypoglycemia_01}
\alias{hypoglycemia_01}
\title{Hypoglycemia-01}
\usage{
hypoglycemia_01(
  df,
  erecord_01_col,
  incident_date_col,
  patient_DOB_col,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  evitals_18_col,
  evitals_23_cl,
  evitals_26_col,
  emedications_03_col,
  eprocedures_03_col,
  ...
)
}
\arguments{
\item{df}{A data frame or tibble containing emergency response records.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column representing the unique record identifier.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> POSIXct or Date column representing the date of the incident.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> POSIXct or Date column representing the patient's date of birth.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column representing the patient's numeric age agnostic of unit.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column representing the patient's age unit ("Years", "Months", "Days", "Hours", or "Minute").}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column containing response type codes.}

\item{esituation_11_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for primary impression fields, containing ICD-10 codes.}

\item{esituation_12_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for secondary impression fields, containing ICD-10 codes.}

\item{evitals_18_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for blood glucose levels.}

\item{evitals_23_cl}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for Glasgow Coma Scale (GCS) scores.}

\item{evitals_26_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for AVPU alertness levels.}

\item{emedications_03_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for administered medications.}

\item{eprocedures_03_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for procedures performed.}

\item{...}{Additional arguments for summarization, passed to the summarize function.}
}
\value{
A tibble summarizing the initial population and filtered subgroups (adult, pediatric, total) with hypoglycemia events.
}
\description{
The \code{hypoglycemia_01} function calculates the NEMSQA measure evaluating how often
hypoglycemic patients with altered mental status receive hypoglycemia treatment.
}
\section{Data Assumptions}{

The \code{df} argument should be a dataframe or tibble with the following assumptions:
\itemize{
\item The data is already loaded.
\item The data has one row per patient/incident and one column for each feature/field.
\item The function will calculate an age in years using the incident date and the patient DOB.
\item The incident date and the patient DOB are Date or POSIXct data types.
\item Any missing values are encoded as \code{NA}, not the "Not Known"/"Not Recorded" text values
or NEMSIS "not value" codes commonly reported by ePCR vendors.
\item The vitals field may be the full list of values for each field or the lowest estimated
eVitals.18 (must include the "Low Blood Glucose" flag) and lowest estimated patient AVPU
in eVitals.2.
\item The function assumes that the primary and secondary impression fields (eSituation.11
and eSituation.12) have the ICD-10 codes in them. The test description can be present,
too, for reference.
\item The function assumes that the eResponse.05 fields have the NEMSIS codes in them,
although text can be also be present for reference.
\item The function assumes that the eMedications_03 and eProcedures_03 fields contain
all medications/procedures and that it contains the text description of the
generic name of the medication. The codes can be included for reference, but will
not be checked. ALL medications and prodcedures are in one field per record, as either
a list column or a comma-separated list.
\item The eSituation.12 (Secondary Impression) field is best as a list column of the secondary
impressions. No joining is done.
\item Any joins to get vitals, etc., will need to be done outside the function.
\item Grouping by specific attributes (e.g., region) can be performed inside this function by
utilizing the \code{.by} argument passed via tidydots (i.e. \code{...}) to \code{dplyr::summarize}.
}
}

\section{Credit}{


This function was developed by (Nicolas Foss, Ed.D., MS)\link{nicolas.foss@hhs.iowa.gov} at the Bureau of Emergency Medical and Trauma
Services, Division of Public Health, Iowa HHS.
}

