% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoglycemia_01_population.R
\name{hypoglycemia_01_population}
\alias{hypoglycemia_01_population}
\title{Hypoglycemia-01 Populations}
\usage{
hypoglycemia_01_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  vitals_table = NULL,
  medications_table = NULL,
  procedures_table = NULL,
  erecord_01_col,
  incident_date_col,
  patient_DOB_col,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  evitals_18_col,
  evitals_23_cl,
  evitals_26_col,
  emedications_03_col,
  eprocedures_03_col
)
}
\arguments{
\item{df}{A data frame or tibble containing emergency response records.}

\item{patient_scene_table}{A data.frame or tibble containing only epatient and escene fields as a fact table.}

\item{response_table}{A data.frame or tibble containing only the eresponse fields needed for this measure's calculations.}

\item{situation_table}{A data.frame or tibble containing only the esituation fields needed for this measure's calculations.}

\item{vitals_table}{A data.frame or tibble containing only the evtials fields needed for this measure's calculations.}

\item{medications_table}{A data.frame or tibble containing only the emedications fields needed for this measure's calculations.}

\item{procedures_table}{A data.frame or tibble containing only the eprocedures fields needed for this measure's calculations.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column representing the EMS record unique identifier.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains the incident date.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains the patient's date of birth.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column representing the patient's numeric age agnostic of unit.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column representing the patient's age unit ("Years", "Months", "Days", "Hours", or "Minute").}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains eResponse.05.}

\item{esituation_11_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains eSituation.11.}

\item{esituation_12_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains all eSituation.12 values as a single comma-separated list.}

\item{evitals_18_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for blood glucose levels.}

\item{evitals_23_cl}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for Glasgow Coma Scale (GCS) scores.}

\item{evitals_26_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for AVPU alertness levels.}

\item{emedications_03_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains all eMedications.03 values as a single comma-separated list.}

\item{eprocedures_03_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for procedures performed.}
}
\value{
#' A list that contains the following:
\itemize{
\item a tibble with counts for each filtering step,
\item a tibble for each population of interest
\item a tibble for the initial population
}
}
\description{
Hypoglycemia-01 Populations
}
\author{
Nicolas Foss, Ed.D., MS
}
