% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoglycemia_01_population.R
\name{hypoglycemia_01_population}
\alias{hypoglycemia_01_population}
\title{Hypoglycemia-01 Populations}
\usage{
hypoglycemia_01_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  vitals_table = NULL,
  medications_table = NULL,
  procedures_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_dob_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  evitals_18_col,
  evitals_23_col,
  evitals_26_col,
  emedications_03_col,
  eprocedures_03_col
)
}
\arguments{
\item{df}{A data frame or tibble containing emergency response records.
Default is \code{NULL}.}

\item{patient_scene_table}{A data.frame or tibble containing at least
epatient and escene fields as a fact table. Default is \code{NULL}.}

\item{response_table}{A data.frame or tibble containing at least the
eresponse fields needed for this measure's calculations. Default is \code{NULL}.}

\item{situation_table}{A data.frame or tibble containing at least the
esituation fields needed for this measure's calculations. Default is
\code{NULL}.}

\item{vitals_table}{A data.frame or tibble containing at least the evitals
fields needed for this measure's calculations. Default is \code{NULL}.}

\item{medications_table}{A data.frame or tibble containing at least the
emedications fields needed for this measure's calculations. Default is
\code{NULL}.}

\item{procedures_table}{A data.frame or tibble containing at least the
eprocedures fields needed for this measure's calculations. Default is
\code{NULL}.}

\item{erecord_01_col}{Column representing the unique record identifier.}

\item{incident_date_col}{Column that contains the incident date. This
defaults to \code{NULL} as it is optional in case not available due to PII
restrictions.}

\item{patient_dob_col}{Column that contains the patient's date of birth. This
defaults to \code{NULL} as it is optional in case not available due to PII
restrictions.}

\item{epatient_15_col}{Column representing the patient's numeric age agnostic
of unit.}

\item{epatient_16_col}{Column representing the patient's age unit ("Years",
"Months", "Days", "Hours", or "Minute").}

\item{eresponse_05_col}{Column containing response type codes.}

\item{esituation_11_col}{Column for primary impression fields, containing
ICD-10 codes.}

\item{esituation_12_col}{Column for secondary impression fields, containing
ICD-10 codes.}

\item{evitals_18_col}{Column for blood glucose levels.}

\item{evitals_23_col}{Column for Glasgow Coma Scale (GCS) scores.}

\item{evitals_26_col}{Column for AVPU alertness levels.}

\item{emedications_03_col}{Column for administered medications.}

\item{eprocedures_03_col}{Column for procedures performed.}
}
\value{
#' A list that contains the following:
\itemize{
\item a tibble with counts for each filtering step,
\item a tibble for each population of interest
\item a tibble for the initial population
}
}
\description{
Filters data down to the target populations for Hypoglycemia-01, and
categorizes records to identify needed information for the calculations.

Identifies key categories related to diabetes/hypoglycemia incidents in an
EMS dataset, specifically focusing on cases where 911 was called for
diabetes/hypoglycemia distress, certain medications were administered, and a
weight is taken. This function segments the data into pediatric populations,
computing the proportion of cases that have a documented weight.
}
\author{
Nicolas Foss, Ed.D., MS
}
