% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoglycemia_01_population.R
\name{hypoglycemia_01_population}
\alias{hypoglycemia_01_population}
\title{Hypoglycemia-01 Populations}
\usage{
hypoglycemia_01_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  vitals_table = NULL,
  medications_table = NULL,
  procedures_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  evitals_18_col,
  evitals_23_cl,
  evitals_26_col,
  emedications_03_col,
  eprocedures_03_col
)
}
\arguments{
\item{df}{A data frame or tibble containing emergency response records.}

\item{patient_scene_table}{A data.frame or tibble containing only epatient and escene fields as a fact table.}

\item{response_table}{A data.frame or tibble containing only the eresponse fields needed for this measure's calculations.}

\item{situation_table}{A data.frame or tibble containing only the esituation fields needed for this measure's calculations.}

\item{vitals_table}{A data.frame or tibble containing only the evtials fields needed for this measure's calculations.}

\item{medications_table}{A data.frame or tibble containing only the emedications fields needed for this measure's calculations.}

\item{procedures_table}{A data.frame or tibble containing only the eprocedures fields needed for this measure's calculations.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column representing the EMS record unique identifier.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains the incident date. This defaults to \code{NULL} as it is optional in case not available due to PII restrictions.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains the patient's date of birth. This defaults to \code{NULL} as it is optional in case not available due to PII restrictions.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column representing the patient's numeric age agnostic of unit.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column representing the patient's age unit ("Years", "Months", "Days", "Hours", or "Minute").}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains eResponse.05.}

\item{esituation_11_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains eSituation.11.}

\item{esituation_12_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains all eSituation.12 values as a single comma-separated list.}

\item{evitals_18_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for blood glucose levels.}

\item{evitals_23_cl}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for Glasgow Coma Scale (GCS) scores.}

\item{evitals_26_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for AVPU alertness levels.}

\item{emedications_03_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains all eMedications.03 values as a single comma-separated list.}

\item{eprocedures_03_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for procedures performed.}
}
\value{
#' A list that contains the following:
\itemize{
\item a tibble with counts for each filtering step,
\item a tibble for each population of interest
\item a tibble for the initial population
}
}
\description{
The \code{hypoglycemia_01_population} filters data down to the target populations for Hypoglycemia-01, and categorizes
records to identify needed information for the calculations.
}
\details{
Identifies key categories related to diabetes/hypoglycemia incidents in an EMS dataset,
specifically focusing on cases where 911 was called for diabetes/hypoglycemia distress,
certain medications were administered, and a weight is taken. This function segments the data
into pediatric populations, computing the proportion of cases that have a documented weight.
}
\section{Data Assumptions}{

The \code{df} argument and each \verb{*_table} argument should be a dataframe or tibble, with the following assumptions:
\itemize{
\item The data are already loaded.
\item The function will calculate an age in years using the incident date and the patient DOB.
\item The incident date and the patient DOB are \code{Date} or \code{POSIXct} data types.
\item Any missing values are encoded as \code{NA}, not the "Not Known"/"Not Recorded" text values
or NEMSIS "not value" codes commonly reported by ePCR vendors.
\item The vitals field may be the full list of values for each field or the lowest estimated
eVitals.18 (must include the "Low Blood Glucose" flag) and lowest estimated patient AVPU
in eVitals.2.
\item The function assumes that the primary and secondary impression fields (eSituation.11
and eSituation.12) have the ICD-10 codes in them. The test description can be present,
too, for reference.
\item The function assumes that the eResponse.05 fields have the NEMSIS codes in them,
although text can be also be present for reference.
\item The function assumes that the eMedications_03 and eProcedures_03 fields contain
all medications/procedures and that it contains the text description of the
generic name of the medication. The codes can be included for reference, but will
not be checked. ALL medications and procedures are in one field per record, as either
a list column or a comma-separated list.
\item The eSituation.12 (Secondary Impression) field is best as a list column of the secondary
impressions. No joining is done.
\item Any joins to get vitals, etc., will need to be done outside the function.
\item Grouping by specific attributes (e.g., region) can be performed inside this function by
utilizing the \code{.by} argument passed via tidydots (i.e. \code{...}) to \code{dplyr::summarize}.
}
}

\author{
Nicolas Foss, Ed.D., MS
}
