% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemsqar-package.R
\docType{package}
\name{nemsqar-package}
\alias{nemsqar}
\alias{nemsqar-package}
\title{nemsqar: National EMS Quality Alliance Measure Calculations}
\description{
nemsqar is an R package designed to automate the calculation of EMS quality metrics as defined by the National EMS Quality Alliance (NEMSQA). With a focus on enhancing EMS service performance through reliable, evidence-based quality measures, nemsqar aims to streamline metric calculation for healthcare providers, EMS agencies, and researchers dedicated to improving patient outcomes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bemts-hhs/nemsqar}
  \item \url{https://bemts-hhs.github.io/nemsqar/}
  \item Report bugs at \url{https://github.com/bemts-hhs/nemsqar/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Foss \email{nicolas.foss@hhs.iowa.gov}

Authors:
\itemize{
  \item Samuel Kordik \email{samuel.kordik@utsouthwestern.edu} (\href{https://orcid.org/0000-0003-4230-1154}{ORCID})
}

Other contributors:
\itemize{
  \item Alyssa Green \email{awhim@nemsqa.org} [contributor]
  \item Iowa Department of Health and Human Services [copyright holder]
}

}
\keyword{internal}
