% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pediatrics_03b.R
\name{pediatrics_03b}
\alias{pediatrics_03b}
\title{Pediatrics-03B Measure Calculation}
\usage{
pediatrics_03b(
  df,
  erecord_01_col,
  incident_date_col,
  patient_DOB_col,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  eexam_01_col,
  eexam_02_col,
  emedications_03_col,
  emedications_04_col,
  ...
)
}
\arguments{
\item{df}{Data frame or tibble containing EMS records.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for unique EMS record identifiers.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Date or POSIXct Column indicating the date of the EMS incident.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Date or POSIXct Column specifying patient date of birth.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column giving the calculated age value.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column giving the provided age unit value.}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column containing the EMS response codes.}

\item{eexam_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column containing documented weight information.}

\item{eexam_02_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Another column for weight documentation, if applicable.}

\item{emedications_03_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column indicating medication administration.}

\item{emedications_04_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column listing medications administered.}

\item{...}{Additional parameters for the \code{dplyr::summarize} output.}
}
\value{
A tibble summarizing results for three population groups (All, Adults, and Peds) with the following columns:

\code{pop}: Population type (All, Adults, Peds).
\code{numerator}: Count of incidents where patient weight was documented.
\code{denominator}: Total count of incidents.
\code{prop}: Proportion of incidents where patient weight was documented.
\code{prop_label}: Proportion formatted as a percentage with a specified number of
decimal places.#'
}
\description{
The \code{pediatrics_03b} function calculates a pediatric metric focused on EMS
responses, specifically targeting responses that involve patients under 18
years of age, where certain weight-based medications were administered. This
function filters EMS data to identify relevant 911 responses and further narrows
down the dataset to cases involving children, calculating the proportion of
cases with documented weight among those where weight-based medications were
administered.
}
\section{Data Assumptions}{

Assume data are already loaded
The data must be a dataframe or tibble that contains emedications.03
as a column where each cell contains all values entered for each respective incident.
This is not a list column but can be comma separated values in each cell, and must contain
all medications for each incident.
emedications.04 is the full list of medication administration routes present in the data
can the function will roll up these values so that the table is distinct with each row being
1 observation, and each column a feature.
this function will calculate an age in years
this function also assumes that rows that are missing any value are NA,
not the not known / not recorded values common to ImageTrend or the value codes
that correspond to "not values".
the function assumes that the eresponse.05 column has the codes in it, text
can be present, too, for reference
the function assumes that edisposition.18 is a list column or a column that has all
text descriptors for additional transport mode descriptors.  These can be separated
by commas or other characters as long as all eresponse.18 values are present
in one cell for each unique erecord.01 value.  Codes can be present
but will be ignored by the function.
for the argument transport_disposition_cols, this argument can receive the unquoted
column names of edisposition.12 and edisposition.30.  One or both can be entered and
the function will evaluate them.  These columns are used to create a \code{transport}
variable that is used to filter the table down furhter.  AS such, these columns
edisposition.12 and edisposition.30 must be list columns that and/or contain all values
from each unique incident entered for each field.  These can be comma separated values
all in one cell to make the table tidy.
the first argument is a dataframe, no joining is done.
any joins to get vitals etc. will need to be done outside the function
Grouping by specific attributes (e.g., region) can be performed inside this function by
utilizing the \code{.by} argument passed via tidydots (i.e. \code{...}) to \code{dplyr::summarize}.
}

\section{Credit}{


This function was developed by (Nicolas Foss, Ed.D., MS)\link{nicolas.foss@hhs.iowa.gov} at the Bureau of Emergency Medical and Trauma
Services, Division of Public Health, Iowa HHS.  (Peter Geissert's)\link{https://www.linkedin.com/in/peter-geissert-ba7607ba} original code base was
used to develop this function.
}

