% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pediatrics_03b_population.R
\name{pediatrics_03b_population}
\alias{pediatrics_03b_population}
\title{Pediatrics-03B Populations}
\usage{
pediatrics_03b_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  exam_table = NULL,
  medications_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  eexam_01_col,
  eexam_02_col,
  emedications_03_col,
  emedications_04_col
)
}
\arguments{
\item{df}{A data frame or tibble containing emergency response records.}

\item{patient_scene_table}{A data.frame or tibble containing only epatient and escene fields as a fact table.}

\item{response_table}{A data.frame or tibble containing only the eresponse fields needed for this measure's calculations.}

\item{exam_table}{A data.frame or tibble containing only the eexam fields needed for this measure's calculations.}

\item{medications_table}{A data.frame or tibble containing only the emedications fields needed for this measure's calculations.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for unique EMS record identifiers.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains the incident date. This defaults to \code{NULL} as it is optional in case not available due to PII restrictions.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column that
contains the patient's date of birth. This defaults to \code{NULL} as it is optional in case not available due to PII restrictions.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column giving the calculated age value.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column giving the provided age unit value.}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column containing the EMS response codes.}

\item{eexam_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column containing documented weight information.}

\item{eexam_02_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Another column for weight documentation, if applicable.}

\item{emedications_03_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column indicating medication administration.}

\item{emedications_04_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column listing medications administered.}
}
\value{
#' A list that contains the following:
\itemize{
\item a tibble with counts for each filtering step,
\item a tibble for each population of interest
\item a tibble for the initial population
}
}
\description{
The \code{pediatrics_03b} function calculates a pediatric metric focused on EMS
responses, specifically targeting responses that involve patients under 18
years of age, where certain weight-based medications were administered. This
function filters EMS data to identify relevant 911 responses and further narrows
down the dataset to cases involving children, calculating the proportion of
cases with documented weight among those where weight-based medications were
administered.
}
\section{Data Assumptions}{

Assume data are already loaded
The data must be a dataframe or tibble that contains emedications.03
as a column where each cell contains all values entered for each respective incident.
This is not a list column but can be comma separated values in each cell, and must contain
all medications for each incident.
emedications.04 is the full list of medication administration routes present in the data
, and the function will deal with these values so that the table is distinct with each row being
1 observation, and each column a feature.
this function will calculate an age in years optionally
this function also assumes that rows that are missing any value are NA,
not the not known / not recorded values common to ImageTrend or the value codes
that correspond to "not values".
the function assumes that the eresponse.05 column has the codes in it, text
can be present, too, for reference
the first argument is a data.frame or the *_table arguments, no joining is done.
any joins to get vitals etc. will need to be done outside the function
Grouping by specific attributes (e.g., region) can be performed inside this function by
utilizing the \code{.by} argument passed via tidydots (i.e. \code{...}) to \code{dplyr::summarize}.
}

\author{
Nicolas Foss, Ed.D., MS
}
