% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respiratory_01_population.R
\name{respiratory_01_population}
\alias{respiratory_01_population}
\title{Respiratory-01 Populations}
\usage{
respiratory_01_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  vitals_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  evitals_12_col,
  evitals_14_col
)
}
\arguments{
\item{df}{A data frame containing incident data with each row representing an
observation.}

\item{patient_scene_table}{A data.frame or tibble containing at least
epatient and escene fields as a fact table.}

\item{response_table}{A data.frame or tibble containing at least the
eresponse fields needed for this measure's calculations.}

\item{situation_table}{A data.frame or tibble containing at least the
esituation fields needed for this measure's calculations.}

\item{vitals_table}{A data.frame or tibble containing at least the evitals
fields needed for this measure's calculations.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Unique Patient ID}

\item{incident_date_col}{Column that contains the incident date. This
defaults to \code{NULL} as it is optional in case not available due to PII
restrictions.}

\item{patient_DOB_col}{Column that contains the patient's date of birth. This
defaults to \code{NULL} as it is optional in case not available due to PII
restrictions.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column giving the
calculated age value.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column giving the
provided age unit value.}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for
911 response codes (e.g., 2205001, 2205003, 2205009).}

\item{esituation_11_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for
primary impression codes related to respiratory distress.}

\item{esituation_12_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for
secondary impression codes related to respiratory distress.}

\item{evitals_12_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for
the first vital sign measurement.}

\item{evitals_14_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for
the second vital sign measurement.}
}
\value{
Returns a data frame summarizing the proportion of cases with
complete vital sign data, divided by population
}
\description{
The \code{respiratory_01_population} function filters and analyzes data related to
emergency 911 respiratory distress incidents, providing the adult, pediatric,
and initial populations. This function uses specific data columns for 911
response codes, primary and secondary impressions, and vital signs to filter
a dataset down to the populations of interest.
}
\author{
Nicolas Foss, Ed.D., MS
}
