% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_01.R
\name{safety_01}
\alias{safety_01}
\title{Safety-01 Calculation}
\usage{
safety_01(
  df,
  incident_date_col,
  patient_DOB_col,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  eresponse_24_col,
  ...
)
}
\arguments{
\item{df}{A data frame or tibble containing EMS data.}

\item{incident_date_col}{Date or POSIXct column indicating the date of the incident.}

\item{patient_DOB_col}{Date or POSIXct column for the patient’s date of birth}

\item{epatient_15_col}{Column containing age.}

\item{epatient_16_col}{Column for age units.}

\item{eresponse_05_col}{Column containing response mode codes (e.g., 911 response codes).}

\item{eresponse_24_col}{Column detailing additional response descriptors as text.}

\item{...}{arguments passed on to summarize.}
}
\value{
Returns a tibble summarizing the Safety-01 metric for all populations, adult, and pediatric groups.
}
\description{
The \code{safety_01} function calculates the proportion of 911 responses where
"lights and sirens" were not used in an EMS dataset. It generates age-based
population summaries, calculating the count and proportion of "lights and
sirens" responses among all incidents, and within adult and pediatric groups.
}
\section{Details}{

This function assumes that:
\itemize{
\item The EMS dataset uses a row-per-observation format.
\item Missing values are NA.
\item The column eresponse.05 contains standardized response codes.
\item The eresponse.24 column contains a list of textual descriptors.
}
}

\section{Features}{

\itemize{
\item Age Calculation: Computes patient age in years by subtracting patient_DOB_col from incident_date_col.
\item Filter for 911 Responses: Filters records based on response codes in eresponse.05.
\item Lights and Sirens Filter: Assigns a value of 1 if "lights and sirens" are not used, and 0 otherwise.
\item Population Grouping: Separates data into adult (age ≥ 18) and pediatric (age < 18) populations.
}
}

\section{Credit}{


This function was developed by (Nicolas Foss, Ed.D., MS)\link{nicolas.foss@hhs.iowa.gov} at the Bureau of Emergency Medical and Trauma
Services, Division of Public Health, Iowa HHS.
}

