% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_01_population.R
\name{safety_01_population}
\alias{safety_01_population}
\title{Safety-01 Populations}
\usage{
safety_01_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  eresponse_24_col
)
}
\arguments{
\item{df}{A data frame or tibble containing EMS data.}

\item{patient_scene_table}{A data frame or tibble containing only epatient and escene fields as a fact table. Default is \code{NULL}.}

\item{response_table}{A data frame or tibble containing only the eresponse fields needed for this measure's calculations. Default is \code{NULL}.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name containing the unique patient record identifier.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Date or POSIXct column indicating the date of the incident.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Date or POSIXct column for the patient’s date of birth}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column containing age.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for age units.}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column containing response mode codes (e.g., 911 response codes).}

\item{eresponse_24_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column detailing additional response descriptors as text.}
}
\value{
#' A list that contains the following:
\itemize{
\item a tibble with counts for each filtering step,
\item a tibble for each population of interest
\item a tibble for the initial population
}
}
\description{
Filters data down to the target populations for Safety-01, and categorizes
records to identify needed information for the calculations.
}
\details{
Identifies key categories related to 911 responses where
"lights and sirens" were not used in an EMS dataset. This function segments the data by
age into adult and pediatric populations.
}
\section{Details}{

This function assumes that:
\itemize{
\item The EMS dataset uses a row-per-observation format.
\item Missing values are NA.
\item The column eresponse.05 contains standardized response codes.
\item The eresponse.24 column contains a list of textual descriptors.
}
}

\section{Features}{

\itemize{
\item Age Calculation: Computes patient age in years by subtracting patient_DOB_col from incident_date_col.
\item Filter for 911 Responses: Filters records based on response codes in eresponse.05.
\item Lights and Sirens Filter: Assigns a value of 1 if "lights and sirens" are not used, and 0 otherwise.
\item Population Grouping: Separates data into adult (age ≥ 18) and pediatric (age < 18) populations.
}
}

\author{
Nicolas Foss, Ed.D., MS
}
