% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_02.R
\name{safety_02}
\alias{safety_02}
\title{Safety-01 Calculation}
\usage{
safety_02(
  df,
  incident_date_col,
  patient_DOB_col,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  edisposition_18_col,
  edisposition_28_col,
  transport_disposition_cols,
  ...
)
}
\arguments{
\item{df}{A data frame where each row is an observation, and each column represents a feature.}

\item{incident_date_col}{Date or POSIXct Unquoted column name representing the date of the incident.}

\item{patient_DOB_col}{Date or POSIXct Unquoted column name for the patient's date of birth.}

\item{epatient_15_col}{Column giving the calculated age value.}

\item{epatient_16_col}{Column giving the provided age unit value.}

\item{eresponse_05_col}{Column giving response codes, identifying 911 responses.}

\item{edisposition_18_col}{Column giving transport mode descriptors, including possible lights-and-sirens indicators.}

\item{edisposition_28_col}{Column giving patient evaluation and care categories for the EMS response.}

\item{transport_disposition_cols}{One or more unquoted column names (such as edisposition.12, edisposition.30) containing transport disposition details.}

\item{...}{Additional arguments for summary calculation, if needed.}
}
\value{
A data frame summarizing the Safety-02 metric with calculated values for three populations:
}
\description{
The \code{safety_02} function calculates the Safety-02 metric, evaluating the
proportion of emergency medical calls involving transport where no lights and
sirens were used. This function categorizes the population into adult and
pediatric groups based on their age, and summarizes results with a total
population count as well.
}
\section{Details}{


Assume data are already loaded
Need to be a table where each row is 1 observation and each column is a feature
or distinct datasets that can be referenced as unique columns
this function will calculate an age in years
this function also assumes that rows that are missing any value are NA,
not the not known / not recorded values common to ImageTrend or the value codes
that correspond to "not values".
the function assumes that the eresponse.05 column has the codes in it, text
can be present, too, for reference
the function assumes that edisposition.18 is a list column or a column that has all
text descriptors for additional transport mode descriptors.  These can be separated
by commas or other characters as long as all eresponse.18 values are present
in one cell for each unique erecord.01 value.  Codes can be present
but will be ignored by the function.
for the argument transport_disposition_cols, this argument can receive the unquoted
column names of edisposition.12 and edisposition.30.  One or both can be entered and
the function will evaluate them.  These columns are used to create a \code{transport}
variable that is used to filter the table down furhter.  AS such, these columns
edisposition.12 and edisposition.30 must be list columns that and/or contain all values
from each unique incident entered for each field.  These can be comma separated values
all in one cell to make the table tidy.
the first argument is a dataframe, no joining is done.
any joins to get vitals etc. will need to be done outside the function
Grouping by specific attributes (e.g., region) can be performed inside this function by
utilizing the \code{.by} argument passed via tidydots (i.e. \code{...}) to \code{dplyr::summarize}.
}

