% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_02_population.R
\name{safety_02_population}
\alias{safety_02_population}
\title{Safety-02 Populations}
\usage{
safety_02_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  disposition_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  edisposition_18_col,
  edisposition_28_col,
  transport_disposition_cols
)
}
\arguments{
\item{df}{A data frame where each row is an observation, and each column represents a feature.}

\item{patient_scene_table}{A data.frame or tibble containing only epatient and escene fields as a fact table.}

\item{response_table}{A data.frame or tibble containing only the eresponse fields needed for this measure's calculations.}

\item{disposition_table}{A data.frame or tibble containing only the edisposition fields needed for this measure's calculations.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Date or POSIXct Unquoted column name representing the date of the incident.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Date or POSIXct Unquoted column name for the patient's date of birth.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column giving the calculated age value.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column giving the provided age unit value.}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column giving response codes, identifying 911 responses.}

\item{edisposition_18_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column giving transport mode descriptors, including possible lights-and-sirens indicators.}

\item{edisposition_28_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column giving patient evaluation and care categories for the EMS response.}

\item{transport_disposition_cols}{<\link[=dplyr_tidy_select]{'tidy-select'}> One or more unquoted column names (such as edisposition.12, edisposition.30) containing transport
disposition details.}
}
\value{
#' A list that contains the following:
\itemize{
\item a tibble with counts for each filtering step,
\item a tibble for each population of interest
\item a tibble for the initial population
}
}
\description{
Filters data down to the target populations for Safety-02, and categorizes
records to identify needed information for the calculations.
}
\details{
Identifies key categories related to a 911 request during which lights and sirens were not used during patient
transport. This function segments the data by age into adult and pediatric populations.
}
\section{Details}{


Assume data are already loaded
Need to be a table where each row is 1 observation and each column is a feature
or distinct datasets that can be referenced as unique columns
this function will calculate an age in years
this function also assumes that rows that are missing any value are NA,
not the not known / not recorded values common to ImageTrend or the value codes
that correspond to "not values".
the function assumes that the eresponse.05 column has the codes in it, text
can be present, too, for reference
the function assumes that edisposition.18 is a list column or a column that has all
text descriptors for additional transport mode descriptors.  These can be separated
by commas or other characters as long as all eresponse.18 values are present
in one cell for each unique erecord.01 value.  Codes can be present
but will be ignored by the function.
for the argument transport_disposition_cols, this argument can receive the unquoted
column names of edisposition.12 and edisposition.30.  One or both can be entered and
the function will evaluate them.  These columns are used to create a \code{transport}
variable that is used to filter the table down furhter.  AS such, these columns
edisposition.12 and edisposition.30 must be list columns that and/or contain all values
from each unique incident entered for each field.  These can be comma separated values
all in one cell to make the table tidy.
the first argument is a dataframe, no joining is done.
any joins to get vitals etc. will need to be done outside the function
Grouping by specific attributes (e.g., region) can be performed inside this function by
utilizing the \code{.by} argument passed via tidydots (i.e. \code{...}) to \code{dplyr::summarize}.
}

\author{
Nicolas Foss, Ed.D., MS
}
