% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_04_population.R
\name{safety_04_population}
\alias{safety_04_population}
\title{Safety-04 Populations}
\usage{
safety_04_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  arrest_table = NULL,
  injury_table = NULL,
  procedures_table = NULL,
  disposition_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_dob_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  earrest_01_col,
  einjury_03_col,
  eprocedures_03_col,
  edisposition_14_col,
  transport_disposition_col
)
}
\arguments{
\item{df}{A data frame or tibble containing EMS data where each row
represents an individual observation.}

\item{patient_scene_table}{A data frame or tibble containing fields from
epatient and escene needed for this measure's calculations.}

\item{response_table}{A data frame or tibble containing fields from eresponse
needed for this measure's calculations.}

\item{arrest_table}{A data frame or tibble containing fields from earrest
needed for this measure's calculations.}

\item{injury_table}{A data frame or tibble containing fields from einjury
needed for this measure's calculations.}

\item{procedures_table}{A data frame or tibble containing fields from
eprocedures needed for this measure's calculations.}

\item{disposition_table}{A data frame or tibble containing fields from
edisposition needed for this measure's calculations.}

\item{erecord_01_col}{The column containing unique record identifiers for
each encounter.}

\item{incident_date_col}{Column that contains the incident date. This
defaults to \code{NULL} as it is optional in case not available due to PII
restrictions.}

\item{patient_dob_col}{Column that contains the patient's date of birth. This
defaults to \code{NULL} as it is optional in case not available due to PII
restrictions.}

\item{epatient_15_col}{Column name indicating the patient age.}

\item{epatient_16_col}{Column name for the unit of age (e.g., "Years,"
"Months").}

\item{eresponse_05_col}{Column containing response transport codes.}

\item{earrest_01_col}{Column with cardiac arrest status information.}

\item{einjury_03_col}{Column describing traumatic injuries, expected as a
list or text-separated entries.}

\item{eprocedures_03_col}{Column listing procedures, assumed to contain
multiple procedure codes/texts in each cell.}

\item{edisposition_14_col}{Column for transport dispositions.}

\item{transport_disposition_col}{Columns for primary and secondary transport
dispositions.}
}
\value{
#' A list that contains the following:
\itemize{
\item a tibble with counts for each filtering step,
\item a tibble for each population of interest
\item a tibble for the initial population
}
}
\description{
Filters data down to the target populations for Safety-04, and categorizes
records to identify needed information for the calculations.

Identifies key categories related to a 911 request or interfacility request
for patients less than 8 years of age during which patients are transported
using a pediatric restraint device. This function segments the data by age
into adult and pediatric populations.
}
\author{
Nicolas Foss, Ed.D., MS
}
