% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seizure_02.R
\name{seizure_02}
\alias{seizure_02}
\title{Seizure-02 Calculation}
\usage{
seizure_02(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  medications_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  emedications_03_col,
  ...
)
}
\arguments{
\item{df}{A data frame where each row is an observation, containing all necessary
columns for analysis.}

\item{patient_scene_table}{A data frame or tibble containing only epatient and escene fields as a fact table. Default is \code{NULL}.}

\item{response_table}{A data frame or tibble containing only the eresponse fields needed for this measure's calculations. Default is \code{NULL}.}

\item{situation_table}{A data.frame or tibble containing only the esituation fields needed for this measure's calculations. Default is \code{NULL}.}

\item{medications_table}{A data.frame or tibble containing only the emedications fields needed for this measure's calculations. Default is \code{NULL}.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column containing unique record identifiers for each encounter.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name representing the incident date as \code{Date} or \code{POSIXct.}}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name representing the patient's date of birth as \code{Date} or \code{POSIXct.}}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for patient age in numeric form.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for age unit (e.g., \code{"Years"} or \code{"Months"}).}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for response codes; "911" call codes are filtered.}

\item{esituation_11_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for primary impressions.}

\item{esituation_12_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for secondary impressions.}

\item{emedications_03_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for medications administered; ideally a list column
or string with comma-separated values.}

\item{...}{Additional arguments passed to \code{dplyr::summarize}.}
}
\value{
A tibble summarizing results for three population groups (All,
Adults, and Peds) with the following columns:
\code{pop}: Population type (All, Adults, or Peds).
\code{numerator}: Count of incidents where beta-agonist medications were administered.
\code{denominator}: Total count of incidents.
\code{prop}: Proportion of incidents involving beta-agonist medications.
\code{prop_label}: Proportion formatted as a percentage with a specified number of
decimal places.
}
\description{
Calculates the NEMSQA Seizure-02 Measure.
}
\details{
Calculates age-based seizure metrics for a dataset. This function filters data
for patients based on incident information, diagnoses, and administered medications
to assess adherence to Seizure-02 metrics.
}
\section{Data Assumptions}{


This function assumes that:

Data are already loaded in a data frame or tibble where each row represents
one observation (e.g., patient), and each column is a feature (field).
Alternatively, data can consist of distinct datasets referenced as unique columns.

Age in years will be calculated using the patient's date of birth and the
incident date. These fields must be valid \code{Date} or \code{POSIXct} data types.

Any missing values in rows are coded as \code{NA} rather than ImageTrend-specific
"not known" or "not recorded" values, or other codes that correspond to
unspecified values.

The primary and secondary impression fields contain ICD-10 codes, although
descriptive text may also be present for reference. These fields are
identified as \code{eSituation.11} for the primary impression and \code{eSituation.12}
for secondary impressions.

The \code{eResponse.05} column contains NEMSIS response codes. Additional descriptive
text may also be present in this field for reference.

The \code{eMedications.03} column holds all medications administered to a patient
during an event, represented as a list column or comma-separated values.
Each entry should use generic medication names, and while RxNORM codes can
be included, they will not be validated by the function.

The \code{eSituation.12} (secondary impressions) field is best prepared as a
comma-separated list within a single string or as a list column containing
each secondary impression.

The first argument to the function is a data frame; any joins needed to
incorporate additional data (e.g., vitals) must be conducted prior to calling
this function.

Grouping by specific attributes (e.g., region) can be performed inside this function by
utilizing the \code{.by} argument passed via tidydots (i.e. \code{...}) to \code{dplyr::summarize}.
}

\section{Practical Tips}{


The first argument is the data.frame prepared as above. No joining is done.
Any joins to get specific data elements will need to be done outside of this function.
}

\section{Value}{

A summarized tibble with counts and proportions of benzodiazepine treatment
among 911 seizure calls, segmented by population group.
}

\author{
Nicolas Foss, Ed.D., MS
}
