% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stroke_01_population.R
\name{stroke_01_population}
\alias{stroke_01_population}
\title{Stroke-01 Populations}
\usage{
stroke_01_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  vitals_table = NULL,
  erecord_01_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  evitals_23_col,
  evitals_26_col,
  evitals_29_col,
  evitals_30_col
)
}
\arguments{
\item{df}{A data frame or tibble containing the dataset. Each row should
represent a unique patient encounter.}

\item{patient_scene_table}{A data frame or tibble containing only epatient
and escene fields as a fact table. Default is \code{NULL}.}

\item{response_table}{A data frame or tibble containing only the eresponse
fields needed for this measure's calculations. Default is \code{NULL}.}

\item{situation_table}{A data.frame or tibble containing only the esituation
fields needed for this measure's calculations. Default is \code{NULL}.}

\item{vitals_table}{A data.frame or tibble containing only the evitals fields
needed for this measure's calculations. Default is \code{NULL}.}

\item{erecord_01_col}{The column containing unique record identifiers for
each encounter.}

\item{eresponse_05_col}{The column containing EMS response codes, which
should include 911 response codes.}

\item{esituation_11_col}{The column containing the primary impression codes
or descriptions related to the situation.}

\item{esituation_12_col}{The column containing secondary impression codes or
descriptions related to the situation.}

\item{evitals_23_col}{The column containing the Glasgow Coma Scale (GCS)
score.}

\item{evitals_26_col}{The column containing the AVPU (alert, verbal, pain,
unresponsive) scale value.}

\item{evitals_29_col}{The column containing the stroke scale score achieved
during assessment.}

\item{evitals_30_col}{The column containing stroke scale type descriptors
(e.g., FAST, NIH, etc.).}
}
\value{
#' A list that contains the following:
\itemize{
\item a tibble with counts for each filtering step,
\item a tibble for each population of interest
\item a tibble for the initial population
}
}
\description{
Filters data down to the target populations for Stroke-01, and categorizes
records to identify needed information for the calculations.

Identifies key categories related to stroke-related incidents in an EMS
dataset, specifically focusing on cases where 911 was called for stroke, and
a stroke scale was administered. .
}
\author{
Nicolas Foss, Ed.D., MS
}
