% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_measure.R
\name{summarize_measure}
\alias{summarize_measure}
\title{Summarize Measure}
\usage{
summarize_measure(data, measure_name, population_name, numerator_col, ...)
}
\arguments{
\item{data}{a dataframe or tibble containing the filtered and calculated fields.}

\item{measure_name}{A string containing the measure description.}

\item{population_name}{A string containing the population description.}

\item{numerator_col}{The tidyselect column containing the numerator.}

\item{...}{(optional) additional arguments}
}
\value{
Summarized information
}
\description{
Calculates measure numerator, denominator, and proportions
for a NEMSQA measure.
}
\details{
Used throughout this package to calculate measure summaries.
}
\author{
Samuel Kordik, BBA, BS
}
