% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncope_01.R
\name{syncope_01}
\alias{syncope_01}
\title{Syncope-01:}
\usage{
syncope_01(
  df,
  incident_date_col,
  patient_DOB_col,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  esituation_09_col,
  esituation_10_col,
  esituation_11_col,
  esituation_12_col,
  evitals_04_col,
  ...
)
}
\arguments{
\item{df}{<\link[=dplyr_tidy_select]{'tidy-select'}> Main data frame containing EMS records.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column containing the incident date, used to calculate age.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column containing the patient's date of birth.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column representing the patient age (numeric).}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for the patient age units (e.g., "Years", "Months").}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column containing response type codes, specifically 911 codes.}

\item{esituation_09_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column with primary symptoms associated with the patient encounter.}

\item{esituation_10_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column with other associated symptoms.}

\item{esituation_11_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for primary impression code.}

\item{esituation_12_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column for secondary impression codes.}

\item{evitals_04_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column with ECG information if available.}

\item{...}{Additional arguments passed to \code{dplyr::summarize} for grouped summaries.}
}
\value{
A tibble summarizing results for three population groups (Adults, and Peds) with the following columns:

\code{pop}: Population type (Adults, Peds).
\code{numerator}: Count of incidents where beta-agonist medications were administered.
\code{denominator}: Total count of incidents.
\code{prop}: Proportion of incidents involving beta-agonist medications.
\code{prop_label}: Proportion formatted as a percentage with a specified number of
decimal places.
}
\description{
The \code{syncope_01} function processes EMS dataset to identify potential syncope (fainting) cases
based on specific criteria and calculates related ECG measures. This function dplyr::filters data for
911 response calls, assesses primary and associated symptoms for syncope, determines age-based
populations (adult and pediatric), and aggregates results by unique patient encounters.
}
\section{Data Assumptions}{


This function assumes that:
\itemize{
\item Data are loaded into a data frame or tibble where each row represents one observation (e.g., patient)
and each column is a distinct feature (field).
\item Missing values in rows are represented as \code{NA}. Any "Not Known" or "Not Recorded" values commonly used
in EMS datasets should be pre-cleaned before use.
\item The \code{eresponse_05_col} column contains NEMSIS response codes, specifically for identifying 911 responses.
\item Symptom and impression columns (\code{esituation_09_col}, \code{esituation_10_col}, \code{esituation_11_col}, and \code{esituation_12_col})
should contain relevant codes or text descriptions that identify syncope and related conditions.  The secondary and associated symptoms
columns need to be columns that contain all values in a comma separated format with all responses in one cell for a unique incident.
These can also be a list column that is unnested before passing to the function.
\item Patient ECG results, if present, are found in \code{evitals_04_col}. Values for this column need to contain all values in a comma
separated format with all responses in one cell for a unique incident. This can also be a list column that is unnested before passing to the
function.
\item Age information is captured across columns \code{epatient_15_col} and \code{epatient_16_col} to distinguish between adults
and pediatric cases.
\item Grouping by specific attributes (e.g., region) can be performed inside this function by
utilizing the \code{.by} argument passed via tidydots (i.e. \code{...}) to \code{dplyr::summarize}.
}

The first argument is the main data frame (\code{df}). No joins are performed; any necessary joins
should be completed prior to calling this function.
}

\section{Practical Tips}{


Ensure data are pre-processed with any missing values coded as \code{NA}. Additionally, ensure that date fields
(e.g., \code{incident_date_col}, \code{patient_DOB_col}) are of type \code{Date} or \code{POSIXct}, as incorrect formatting will
trigger an error message.
}

\author{
Nicolas Foss, Ed.D., MS
}
