% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi_01.R
\name{tbi_01}
\alias{tbi_01}
\title{TBI Screening Function}
\usage{
tbi_01(
  df,
  erecord_01_col,
  incident_date_col,
  patient_DOB_col,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  evitals_23_col,
  evitals_26_col,
  transport_disposition_col,
  evitals_12_col,
  evitals_16_col,
  evitals_06_col,
  ...
)
}
\arguments{
\item{df}{A data frame or tibble containing the patient data.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} with the patient’s unique record ID.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} with the incident date.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} with the patient's date of birth.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} with the patient’s age value.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} with the patient’s age unit (e.g., years, months).}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} with response codes for the type of EMS call.}

\item{esituation_11_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} with the primary provider impression.}

\item{esituation_12_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} with the secondary provider impression.}

\item{evitals_23_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} with Glasgow Coma Scale (GCS) scores.}

\item{evitals_26_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} with AVPU (alert, verbal, painful, unresponsive) values.}

\item{transport_disposition_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} with the transport disposition.}

\item{evitals_12_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} with pulse oximetry values.}

\item{evitals_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} with ETCO2 values.}

\item{evitals_06_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} with systolic blood pressure (SBP) values.}

\item{...}{Additional parameters passed to \code{\link[dplyr]{mutate}} or other dplyr functions.}
}
\value{
A tibble summarizing results for three population groups (Adults, and Peds) with the following columns:

\code{pop}: Population type (Adults, Peds).
\code{numerator}: Count of incidents where SP02, ETCO2, and SBP were all measured.
\code{denominator}: Total count of incidents.
\code{prop}: Proportion of incidents where SP02, ETCO2, and SBP were all measured.
\code{prop_label}: Proportion formatted as a percentage with a specified number of
decimal places.
}
\description{
This function screens for potential traumatic brain injury (TBI) cases based on specific criteria
in a patient dataset. It produces a subset of the data with calculated variables for TBI identification.
}
\details{
The function performs data screening and processing to identify potential TBI cases by applying the following steps:
\itemize{
\item Ensures required columns are present and correctly formatted (e.g., date columns).
\item Creates unique IDs based on patient identifiers and incident dates.
\item Filters and selects patients who meet criteria for TBI indicators (e.g., Glasgow Coma Scale < 15, specific AVPU values, and TBI injury codes).
\item Generates derived variables for age classification, system-calculated age check, and other indicators.
\item Returns a filtered data frame suitable for TBI case identification.
}
}
\note{
This function relies on columns formatted as \code{Date} or \code{POSIXct} for dates and expects numeric or categorical
values for other health indicators.
}
\author{
Nicolas Foss, Ed.D., MS
}
