% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi_01_population.R
\name{tbi_01_population}
\alias{tbi_01_population}
\title{TBI-01 Populations}
\usage{
tbi_01_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  disposition_table = NULL,
  vitals_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_dob_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  transport_disposition_col,
  evitals_06_col,
  evitals_12_col,
  evitals_16_col,
  evitals_23_col,
  evitals_26_col
)
}
\arguments{
\item{df}{A data frame or tibble containing the patient data.}

\item{patient_scene_table}{A data frame or tibble containing only epatient
and escene fields as a fact table. Default is \code{NULL}.}

\item{response_table}{A data frame or tibble containing only the eresponse
fields needed for this measure's calculations. Default is \code{NULL}.}

\item{situation_table}{A data.frame or tibble containing only the esituation
fields needed for this measure's calculations. Default is \code{NULL}.}

\item{disposition_table}{A data.frame or tibble containing only the
edisposition fields needed for this measure's calculations. Default is
\code{NULL}.}

\item{vitals_table}{A data.frame or tibble containing only the evitals fields
needed for this measure's calculations. Default is \code{NULL}.}

\item{erecord_01_col}{Column name in df with the patient’s unique record ID.}

\item{incident_date_col}{Column that contains the incident date. This
defaults to \code{NULL} as it is optional in case not available due to PII
restrictions.}

\item{patient_dob_col}{Column that contains the patient's date of birth. This
defaults to \code{NULL} as it is optional in case not available due to PII
restrictions.}

\item{epatient_15_col}{Column name in df with the patient’s age value.}

\item{epatient_16_col}{Column name in df with the patient’s age unit (e.g.,
years, months).}

\item{eresponse_05_col}{Column name in df with response codes for the type of
EMS call.}

\item{esituation_11_col}{Column name in df with the primary provider
impression.}

\item{esituation_12_col}{Column name in df with the secondary provider
impression.}

\item{transport_disposition_col}{Column name in df with the transport
disposition.}

\item{evitals_06_col}{Column name in df with systolic blood pressure (SBP)
values.}

\item{evitals_12_col}{Column name in df with pulse oximetry values.}

\item{evitals_16_col}{Column name in df with ETCO2 values.}

\item{evitals_23_col}{Column name in df with Glasgow Coma Scale (GCS) scores.}

\item{evitals_26_col}{Column name in df with AVPU (alert, verbal, painful,
unresponsive) values.}
}
\value{
#' A list that contains the following:
\itemize{
\item a tibble with counts for each filtering step,
\item a tibble for each population of interest
\item a tibble for the initial population
}
}
\description{
This function screens for potential traumatic brain injury (TBI) cases based
on specific criteria in a patient dataset. It produces a subset of the data
with calculated variables for TBI identification.
}
\author{
Nicolas Foss, Ed.D., MS
}
