% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cleaner.R
\name{theme_cleaner}
\alias{theme_cleaner}
\title{Theme Cleaner}
\usage{
theme_cleaner(...)
}
\arguments{
\item{...}{Passed on to theme_clean()}
}
\value{
ggplot2 theme.
}
\description{
Temporary Theme for theme_cleaner function
}
