% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trauma_01.R
\name{trauma_01}
\alias{trauma_01}
\title{Trauma-01 Calculation}
\usage{
trauma_01(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  disposition_table = NULL,
  vitals_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  esituation_02_col,
  eresponse_05_col,
  evitals_23_col,
  evitals_26_col,
  evitals_27_col,
  edisposition_28_col,
  transport_disposition_col,
  ...
)
}
\arguments{
\item{df}{A data frame or tibble containing EMS records. Default is \code{NULL}.}

\item{patient_scene_table}{A data frame or tibble containing only epatient and escene fields as a fact table. Default is \code{NULL}.}

\item{response_table}{A data frame or tibble containing only the eresponse fields needed for this measure's calculations. Default is \code{NULL}.}

\item{situation_table}{A data.frame or tibble containing only the esituation fields needed for this measure's calculations. Default is \code{NULL}.}

\item{disposition_table}{A data.frame or tibble containing only the edisposition fields needed for this measure's calculations. Default is \code{NULL}.}

\item{vitals_table}{A data.frame or tibble containing only the evitals fields needed for this measure's calculations. Default is \code{NULL}.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name representing the EMS record ID.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the incident date. Default is \code{NULL}.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the patient's date of birth. Default is \code{NULL}.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the patient's age in numeric format.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the unit of age (e.g., "Years", "Months").}

\item{esituation_02_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name indicating if the situation involved an injury.}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the type of EMS response (e.g., 911 call).}

\item{evitals_23_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the Glasgow Coma Scale (GCS) total score.}

\item{evitals_26_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for AVPU (Alert, Voice, Pain, Unresponsive) status.}

\item{evitals_27_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the pain scale assessment.}

\item{edisposition_28_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for patient care disposition details.}

\item{transport_disposition_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for transport disposition details.}

\item{...}{Additional arguments passed to the \code{summarize_measure} function for custom summarization.}
}
\value{
A tibble summarizing results for three population groups (All, Adults, and Peds) with the following columns:

\code{pop}: Population type (All, Adults, Peds).
\code{numerator}: Count of incidents where a pain scale was administered.
\code{denominator}: Total count of incidents.
\code{prop}: Proportion of incidents where a pain scale was administered.
\code{prop_label}: Proportion formatted as a percentage with a specified number of
decimal places.
}
\description{
This function processes EMS data to calculate the Trauma-01 performance measure,
which evaluates the percentage of trauma patients assessed for pain using a numeric scale.
The function filters and summarizes the data based on specified inclusion criteria.
}
\section{Data Assumptions}{

\itemize{
\item \code{evitals_23_col} is the highest GCS total score or all GCS total scores.
\item \code{evitals_26_col} contains all the AVPU responses for all records.
\item \code{evitals_27_col} is the initial pain scale score or all pain scale scores.
}
}

\author{
Nicolas Foss, Ed.D., MS
}
