% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trauma_01.R
\name{trauma_01}
\alias{trauma_01}
\title{Trauma-01}
\usage{
trauma_01(
  df,
  erecord_01_col,
  incident_date_col,
  patient_DOB_col,
  epatient_15_col,
  epatient_16_col,
  esituation_02_col,
  evitals_23_col,
  evitals_26_col,
  eresponse_05_col,
  edisposition_28_col,
  transport_disposition_col,
  evitals_27_col,
  ...
)
}
\arguments{
\item{df}{<\link[=dplyr_tidy_select]{'tidy-select'}> A data frame or tibble containing EMS records.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name representing the EMS record ID.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the incident date.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the patient's date of birth.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the patient's age in numeric format.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the unit of age (e.g., "Years", "Months").}

\item{esituation_02_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name indicating if the situation involved an injury.}

\item{evitals_23_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the Glasgow Coma Scale (GCS) total score.}

\item{evitals_26_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for AVPU (Alert, Voice, Pain, Unresponsive) status.}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the type of EMS response (e.g., 911 call).}

\item{edisposition_28_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for patient care disposition details.}

\item{transport_disposition_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for transport disposition details.}

\item{evitals_27_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the pain scale assessment.}

\item{...}{Additional arguments passed to the \code{summarize_measure} function for custom summarization.
#'}
}
\value{
A tibble summarizing results for three population groups (All, Adults, and Peds) with the following columns:

\code{pop}: Population type (All, Adults, Peds).
\code{numerator}: Count of incidents where a pain scale was administered.
\code{denominator}: Total count of incidents.
\code{prop}: Proportion of incidents where a pain scale was administered.
\code{prop_label}: Proportion formatted as a percentage with a specified number of
decimal places.
}
\description{
This function processes EMS data to calculate the Trauma-01 performance measure,
which evaluates the percentage of trauma patients assessed for pain using a numeric scale.
The function filters and summarizes the data based on specified inclusion criteria.
}
\details{
The function performs the following steps:
\itemize{
\item Validates input data for proper formats and types.
\item Creates unique IDs for patient incidents to maintain row distinctness.
\item Filters records based on specific criteria, including injury status, GCS or AVPU values,
response type (911), patient care disposition, and transport type.
\item Separately identifies adult and pediatric populations based on system and calculated age.
\item Summarizes the Trauma-01 measure for the entire population, adults, and pediatrics.
}
}
