% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trauma_01_population.R
\name{trauma_01_population}
\alias{trauma_01_population}
\title{Trauma-01 Population}
\usage{
trauma_01_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  disposition_table = NULL,
  vitals_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  esituation_02_col,
  eresponse_05_col,
  evitals_23_col,
  evitals_26_col,
  evitals_27_col,
  edisposition_28_col,
  transport_disposition_col
)
}
\arguments{
\item{df}{A data frame or tibble containing EMS records. Default is \code{NULL}.}

\item{patient_scene_table}{A data frame or tibble containing only epatient and escene fields as a fact table. Default is \code{NULL}.}

\item{response_table}{A data frame or tibble containing only the eresponse fields needed for this measure's calculations. Default is \code{NULL}.}

\item{situation_table}{A data.frame or tibble containing only the esituation fields needed for this measure's calculations. Default is \code{NULL}.}

\item{disposition_table}{A data.frame or tibble containing only the edisposition fields needed for this measure's calculations. Default is \code{NULL}.}

\item{vitals_table}{A data.frame or tibble containing only the evitals fields needed for this measure's calculations. Default is \code{NULL}.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name representing the EMS record ID.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the incident date. Default is \code{NULL}.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the patient's date of birth. Default is \code{NULL}.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the patient's age in numeric format.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the unit of age (e.g., "Years", "Months").}

\item{esituation_02_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name indicating if the situation involved an injury.}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the type of EMS response (e.g., 911 call).}

\item{evitals_23_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the Glasgow Coma Scale (GCS) total score.}

\item{evitals_26_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for AVPU (Alert, Voice, Pain, Unresponsive) status.}

\item{evitals_27_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for the pain scale assessment.}

\item{edisposition_28_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for patient care disposition details.}

\item{transport_disposition_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name for transport disposition details.}
}
\value{
#' A list that contains the following:
\itemize{
\item a tibble with counts for each filtering step,
\item a tibble for each population of interest
\item a tibble for the initial population
}
}
\description{
Filters data down to the target populations for Trauma-08, and categorizes
records to identify needed information for the calculations.
}
\details{
Identifies key categories to records that are 911 requests for patients with injury
who were assessed for pain based on specific criteria and calculates related ECG measures.
This function segments the data by age into adult and pediatric populations.
}
\section{Data Assumptions}{

\itemize{
\item \code{evitals_23_col} is the highest GCS total score or all GCS total scores.
\item \code{evitals_26_col} contains all the AVPU responses for all records.
\item \code{evitals_27_col} is the initial pain scale score or all pain scale scores.
}
}

\author{
Nicolas Foss, Ed.D., MS
}
