% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trauma_03.R
\name{trauma_03}
\alias{trauma_03}
\title{Trauma-03 Calculation}
\usage{
trauma_03(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  disposition_table = NULL,
  vitals_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  esituation_02_col,
  eresponse_05_col,
  edisposition_28_col,
  transport_disposition_col,
  evitals_01_col,
  evitals_27_col = NULL,
  evitals_27_initial_col = NULL,
  evitals_27_last_col = NULL,
  ...
)
}
\arguments{
\item{df}{A data frame or tibble containing EMS data with all relevant columns. Default is \code{NULL}.}

\item{patient_scene_table}{A data frame or tibble containing only epatient and escene fields as a fact table. Default is \code{NULL}.}

\item{response_table}{A data frame or tibble containing only the eresponse fields needed for this measure's calculations. Default is \code{NULL}.}

\item{situation_table}{A data frame or tibble containing only the esituation fields needed for this measure's calculations. Default is \code{NULL}.}

\item{disposition_table}{A data frame or tibble containing only the edisposition fields needed for this measure's calculations. Default is \code{NULL}.}

\item{vitals_table}{A data frame or tibble containing only the evitals fields needed for this measure's calculations. Default is \code{NULL}.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column representing the EMS record unique identifier.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column indicating the incident date. Must be of class \code{Date} or similar. Default is \code{NULL}.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column representing the patient's date of birth. Must be of class \code{Date} or similar. Default is \code{NULL}.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for patient age numeric value.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for patient age unit (e.g., "Years", "Months").}

\item{esituation_02_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column containing information on the presence of injury.}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column representing the 911 response type.}

\item{edisposition_28_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for patient care disposition details.}

\item{transport_disposition_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for patient transport disposition.}

\item{evitals_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for the time of pain scale measurement.}

\item{evitals_27_initial_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for the initial pain scale score.}

\item{evitals_27_last_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for the last pain scale score.}

\item{...}{Additional arguments passed to helper functions for further customization.}

\item{evitals_27_sortorder_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for the sort order of pain scale measurement.}
}
\value{
A tibble summarizing results for three population groups (All, Adults, and Peds) with the following columns:

\code{pop}: Population type (All, Adults, Peds).
\code{numerator}: Count of incidents where there was a reduction in patient pain.
\code{denominator}: Total count of incidents.
\code{prop}: Proportion of incidents where there was a reduction in patient pain.
\code{prop_label}: Proportion formatted as a percentage with a specified number of
decimal places.
}
\description{
This function calculates the "Trauma-03" measure, which evaluates pain scale reassessment for trauma patients, using a comprehensive data frame with EMS records. The function processes input data to create both fact and dimension tables, identifies eligible patients, and summarizes results for adult and pediatric populations.
}
\section{Data Assumptions}{

\itemize{
\item If a user passes the \code{evitals_27_initial_col} and \code{evitals_27_last_col}, the function
will utilize those values to do calculations for the numerator.
\item In the case that those values are \code{NULL}, the \code{evitals_27_col} must be passed
and the function will do its best to find the first and last pain scores based on
the \code{evitals_01_col} values.
\item \code{evitals_27_initial_col} is the first pain scale score.
\item \code{evitals_27_last_col} is the final pain scale score.
\item \code{evitals_01_col} contains all vital measurement times, even those not corresponding
to the pain scale.
}
}

\author{
Nicolas Foss, Ed.D., MS
}
