% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trauma_03_population.R
\name{trauma_03_population}
\alias{trauma_03_population}
\title{Trauma-03 Populations}
\usage{
trauma_03_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  disposition_table = NULL,
  vitals_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  esituation_02_col,
  eresponse_05_col,
  evitals_01_col,
  evitals_27_col = NULL,
  evitals_27_initial_col = NULL,
  evitals_27_last_col = NULL,
  edisposition_28_col,
  transport_disposition_col
)
}
\arguments{
\item{df}{A data frame or tibble containing EMS data with all relevant columns. Default is \code{NULL}.}

\item{patient_scene_table}{A data frame or tibble containing only epatient and escene fields as a fact table. Default is \code{NULL}.}

\item{response_table}{A data frame or tibble containing only the eresponse fields needed for this measure's calculations. Default is \code{NULL}.}

\item{situation_table}{A data frame or tibble containing only the esituation fields needed for this measure's calculations. Default is \code{NULL}.}

\item{disposition_table}{A data frame or tibble containing only the edisposition fields needed for this measure's calculations. Default is \code{NULL}.}

\item{vitals_table}{A data frame or tibble containing only the evitals fields needed for this measure's calculations. Default is \code{NULL}.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column representing the EMS record unique identifier.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column indicating the incident date. Must be of class \code{Date} or similar. Default is \code{NULL}.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column representing the patient's date of birth. Must be of class \code{Date} or similar. Default is \code{NULL}.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for patient age numeric value.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for patient age unit (e.g., "Years", "Months").}

\item{esituation_02_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column containing information on the presence of injury.}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column representing the 911 response type.}

\item{evitals_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for the time of pain scale measurement.}

\item{evitals_27_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for the pain scale score. Default is \code{NULL}.}

\item{evitals_27_initial_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for the initial pain scale score. Default is \code{NULL}.}

\item{evitals_27_last_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for the last pain scale score. Default is \code{NULL}.}

\item{edisposition_28_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for patient care disposition details.}

\item{transport_disposition_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for patient transport disposition.}
}
\value{
#' A list that contains the following:
\itemize{
\item a tibble with counts for each filtering step,
\item a tibble for each population of interest
\item a tibble for the initial population
}
}
\description{
Filters data down to the target populations for Trauma-08, and categorizes
records to identify needed information for the calculations.
}
\details{
Identifies key categories to records that are 911 request for patients whose pain score was
lowered during the EMS encounter. based on specific criteria and calculates related ECG measures.
This function segments the data by age into adult and pediatric populations.
}
\section{Data Assumptions}{

\itemize{
\item If a user passes the \code{evitals_27_initial_col} and \code{evitals_27_last_col}, the function
will utilize those values to do calculations for the numerator.
\item In the case that those values are \code{NULL}, the \code{evitals_27_col} must be passed
and the function will do its best to find the first and last pain scores based on
the \code{evitals_01_col} values.
\item \code{evitals_27_initial_col} is the first pain scale score.
\item \code{evitals_27_last_col} is the final pain scale score.
\item \code{evitals_01_col} contains all vital measurement times, even those not corresponding
to the pain scale.
}
}

\author{
Nicolas Foss, Ed.D., MS
}
