% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trauma_08.R
\name{trauma_08}
\alias{trauma_08}
\title{Trauma-08 Measure Calculation}
\usage{
trauma_08(
  df,
  erecord_01_col,
  incident_date_col,
  patient_DOB_col,
  epatient_15_col,
  epatient_16_col,
  esituation_02_col,
  eresponse_05_col,
  transport_disposition_col,
  evitals_14_col,
  evitals_06_col,
  evitals_23_col,
  ...
)
}
\arguments{
\item{df}{A data frame or tibble containing the dataset to analyze.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column specifying unique patient records.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column indicating the incident date. Must be of class \code{Date} or \code{POSIXct}.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column indicating the patient's date of birth. Must be of class \code{Date} or \code{POSIXct}.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column indicating the patient’s age in numeric form.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column specifying the unit of patient age (e.g., "Years", "Days").}

\item{esituation_02_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing information about the nature of the patient’s condition (e.g., injury type).}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column specifying the type of response (e.g., 911 codes).}

\item{transport_disposition_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column specifying transport disposition for the patient.}

\item{evitals_14_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing respiratory rate data from initial vital signs.}

\item{evitals_06_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing systolic blood pressure (SBP) data from initial vital signs.}

\item{evitals_23_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing total Glasgow Coma Scale (GCS) scores from initial vital signs.}

\item{...}{Additional arguments passed to the \code{summarize_measure} function.}
}
\value{
A tibble summarizing results for three population groups (All, Adults, and Peds) with the following columns:

\code{pop}: Population type (All, Adults, Peds).
\code{numerator}: Count of incidents where the respiratory rate, SBP, and GCS vitals were taken.
\code{denominator}: Total count of incidents.
\code{prop}: Proportion of incidents where the respiratory rate, SBP, and GCS vitals were taken.
\code{prop_label}: Proportion formatted as a percentage with a specified number of
decimal places.
}
\description{
This function calculates the Trauma-08 measure, which evaluates the completeness of pain scale documentation for patients experiencing traumatic injury. It determines the total population, adult population, and pediatric population meeting the criteria for the Trauma-08 measure. \strong{Note:} This function assumes the input dataset contains the \emph{initial} vital signs for respiratory rate, systolic blood pressure (SBP), and total Glasgow Coma Scale (GCS) score, respectively.
}
\details{
The function performs the following steps:
\itemize{
\item Validates input data for proper formats and types.
\item Creates unique IDs for patient incidents to maintain row distinctness.
\item Filters records based on specific criteria, including injury status, response type (911),
and transport type.
\item Separately identifies adult and pediatric populations based on system and calculated age.
\item Summarizes the Trauma-08 measure for the entire population, adults, and pediatrics.
}
}
\note{
\itemize{
\item Ensure the input dataset contains initial vital signs for each required vital signs.
\item Not values should not be used, values missing a response should be blank so R interprets those as \code{NA} values.
\item Date columns (\code{incident_date_col} and \code{patient_DOB_col}) must be properly formatted before calling this function.
}
}
\section{Features}{

\itemize{
\item Handles missing or invalid date formats with error messaging.
\item Incorporates quasiquotation for flexible column referencing.
\item Creates reusable dimension tables for efficient filtering and summarization.
}
}

\author{
Nicolas Foss, Ed.D., MS
}
