% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trauma_14.R
\name{trauma_14}
\alias{trauma_14}
\title{Trauma-14}
\usage{
trauma_14(
  df,
  erecord_01_col,
  incident_date_col,
  patient_DOB_col,
  epatient_15_col,
  epatient_16_col,
  esituation_02_col,
  eresponse_05_col,
  transport_disposition_col,
  evitals_21_col,
  evitals_14_col,
  eexam_23_col,
  eexam_25_col,
  evitals_15_col,
  eprocedures_03_col,
  evitals_12_col,
  evitals_06_col,
  evitals_10_col,
  einjury_03_col,
  eexam_16_col,
  eexam_20_col,
  einjury_04_col,
  einjury_09_col,
  eresponse_10_col,
  einjury_01_col,
  ...
)
}
\arguments{
\item{df}{A data frame or tibble containing EMS data with all relevant columns.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column representing the EMS record unique identifier.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column indicating the incident date. Must be of class \code{Date} or similar.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column representing the patient's date of birth. Must be of class \code{Date} or similar.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for patient age numeric value.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for patient age unit (e.g., "Years", "Months").}

\item{esituation_02_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column containing information on the presence of injury.}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column representing the 911 response type.}

\item{transport_disposition_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for patient transport disposition.}

\item{evitals_21_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing Glasgow Coma Scale (GCS) Motor values.}

\item{evitals_14_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing capillary refill information.}

\item{eexam_23_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing lung assessment details.}

\item{eexam_25_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing chest assessment details.}

\item{evitals_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing respiratory effort values.}

\item{eprocedures_03_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing airway management or tourniquet usage details.}

\item{evitals_12_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing pulse oximetry values.}

\item{evitals_06_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing systolic blood pressure (SBP) values.}

\item{evitals_10_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing heart rate values.}

\item{einjury_03_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing trauma triage steps 1 and 2 information.}

\item{eexam_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing extremities assessment details.}

\item{eexam_20_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing neurological assessment details.}

\item{einjury_04_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing trauma triage steps 3 and 4 information.}

\item{einjury_09_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing fall height information.}

\item{eresponse_10_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing scene delay information.}

\item{einjury_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing injury cause details.}

\item{...}{Additional arguments passed to helper functions for further customization.}

\item{edisposition_24_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing pre-hospital trauma alert information.}
}
\value{
A tibble summarizing results for three age groups (< 10 yrs, 10–65 yrs, and >= 65 yrs) with the following columns:

\code{pop}: Population type (< 10 yrs, 10–65 yrs, >= 65 yrs).
\code{numerator}: Count of incidents where a pre-hospital trauma alert was called.
\code{denominator}: Total count of incidents.
\code{prop}: Proportion of incidents where a pre-hospital trauma alert was called.
\code{prop_label}: Proportion formatted as a percentage with a specified number of decimal places.
}
\description{
This function processes EMS data to generate a set of binary variables indicating whether specific trauma triage criteria are met. The output #' is a data frame enriched with these indicators for further analysis.  The final outcome is whether or not the EMS record documents the use of #' a pre-hospital trauma activation.
}
\details{
The function generates individual vectors of unique IDs for each trauma triage criterion and assigns these to binary variables in the final data frame. This allows filtering and calculation of metrics such as numerators for performance measures.
}
\note{
This function uses \code{rlang}, \code{lubridate}, \code{dplyr}, and \code{tidyr} packages for data processing. Ensure the data frame contains valid date formats and expected column names.
}
\section{Features}{

\itemize{
\item Handles missing or invalid date formats with error messaging.
\item Incorporates quasiquotation for flexible column referencing.
\item Creates reusable dimension tables for efficient filtering and summarization.
}
}

\author{
Nicolas Foss, Ed.D., MS
}
