% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trauma_14_population.R
\name{trauma_14_population}
\alias{trauma_14_population}
\title{Trauma-14}
\usage{
trauma_14_population(
  df = NULL,
  patient_scene_table = NULL,
  situation_table = NULL,
  response_table = NULL,
  disposition_table = NULL,
  vitals_table = NULL,
  exam_table = NULL,
  procedures_table = NULL,
  injury_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  esituation_02_col,
  eresponse_05_col,
  eresponse_10_col,
  transport_disposition_col,
  edisposition_23_col,
  evitals_06_col,
  evitals_10_col,
  evitals_12_col,
  evitals_14_col,
  evitals_15_col,
  evitals_21_col,
  eexam_16_col,
  eexam_20_col,
  eexam_23_col,
  eexam_25_col,
  eprocedures_03_col,
  einjury_01_col,
  einjury_03_col,
  einjury_04_col,
  einjury_09_col,
  ...
)
}
\arguments{
\item{df}{A data frame or tibble containing EMS data with all relevant columns.}

\item{patient_scene_table}{A data frame or tibble containing fields from epatient and escene needed for this measure's calculations.}

\item{situation_table}{A data frame or tibble containing fields from esituation needed for this measure's calculations.}

\item{response_table}{A data frame or tibble containing fields from eresponse needed for this measure's calculations.}

\item{disposition_table}{A data frame or tibble containing fields from edisposition needed for this measure's calculations.}

\item{vitals_table}{A data frame or tibble containing fields from evitals needed for this measure's calculations.}

\item{exam_table}{A data frame or tibble containing fields from eexam needed for this measure's calculations.}

\item{procedures_table}{A data frame or tibble containing fields from eprocedures needed for this measure's calculations.}

\item{injury_table}{A data frame or tibble containing fields from einjury needed for this measure's calculations.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column representing the EMS record unique identifier.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column indicating the incident date. Must be of class \code{Date} or similar.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column representing the patient's date of birth. Must be of class \code{Date} or similar.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for patient age numeric value.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for patient age unit (e.g., "Years", "Months").}

\item{esituation_02_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column containing information on the presence of injury.}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column representing the 911 response type.}

\item{eresponse_10_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing scene delay information.}

\item{transport_disposition_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> The column for patient transport disposition.}

\item{edisposition_23_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing pre-hospital trauma alert information.}

\item{evitals_06_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing systolic blood pressure (SBP) values.}

\item{evitals_10_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing heart rate values.}

\item{evitals_12_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing pulse oximetry values.}

\item{evitals_14_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing capillary refill information.}

\item{evitals_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing respiratory effort values.}

\item{evitals_21_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing Glasgow Coma Scale (GCS) Motor values.}

\item{eexam_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing extremities assessment details.}

\item{eexam_20_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing neurological assessment details.}

\item{eexam_23_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing lung assessment details.}

\item{eexam_25_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing chest assessment details.}

\item{eprocedures_03_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing airway management or tourniquet usage details.}

\item{einjury_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing injury cause details.}

\item{einjury_03_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing trauma triage steps 1 and 2 information.}

\item{einjury_04_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing trauma triage steps 3 and 4 information.}

\item{einjury_09_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> Column name in \code{df} containing fall height information.}

\item{...}{Additional arguments passed to helper functions for further customization.}
}
\value{
#' A list that contains the following:
\itemize{
\item a tibble with counts for each filtering step,
\item a tibble for each population of interest
\item a tibble for the initial population
}
}
\description{
This function processes EMS data to generate the population needed to calculated the Trauma-14 NEMSQA measure.
}
\section{Data Assumptions}{

\itemize{
\item All vitals columns contain the full list of values entered for each record and columns.
\item Each eexam field is a list column containing all values entered per record
that are also comma (or otherwise) separated.
\item The trauma triage fields are list columns containing all values entered per record
that are also comma (or otherwise) separated.
\item \code{einjury_01_col} is a list column containing all values entered per record
that are also comma (or otherwise) separated.
\item \code{eresponse_10_col} is a list column containing all values entered per record
that are also comma (or otherwise) separated.
\item \code{eprocedures_03_col} is a list column containing all values entered per record
that are also comma (or otherwise) separated.
}
}

\section{Features}{

\itemize{
\item Handles missing or invalid date formats with error messaging.
\item Incorporates quasiquotation for flexible column referencing.
\item Creates reusable dimension tables for efficient filtering and summarization.
}
}

\author{
Nicolas Foss, Ed.D., MS
}
