% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttr_01.R
\name{ttr_01}
\alias{ttr_01}
\title{TTR-01 Calculation}
\usage{
ttr_01(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  disposition_table = NULL,
  vitals_table = NULL,
  arrest_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  transport_disposition_col,
  earrest_01_col,
  evitals_06_col,
  evitals_07_col,
  evitals_10_col,
  evitals_12_col,
  evitals_14_col,
  evitals_23_col,
  evitals_26_col,
  ...
)
}
\arguments{
\item{df}{A data frame or tibble containing the dataset to analyze. Default is \code{NULL}.}

\item{patient_scene_table}{A data frame or tibble containing only epatient and escene fields as a fact table. Default is \code{NULL}.}

\item{response_table}{A data frame or tibble containing only the eresponse fields needed for this measure's calculations. Default is \code{NULL}.}

\item{disposition_table}{A data frame or tibble containing only the edisposition fields needed for this measure's calculations. Default is \code{NULL}.}

\item{vitals_table}{A data frame or tibble containing only the evitals fields needed for this measure's calculations. Default is \code{NULL}.}

\item{arrest_table}{A data frame or tibble containing only the earrest fields needed for this measure's calculations. Default is \code{NULL}.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column specifying unique patient records.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column indicating the incident date. Must be of class \code{Date} or \code{POSIXct}.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column indicating the patient's date of birth. Must be of class \code{Date} or \code{POSIXct}.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column indicating the patient’s age in numeric form.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column specifying the unit of patient age (e.g., "Years", "Days").}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column specifying the type of response (e.g., 911 codes).}

\item{transport_disposition_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column specifying transport disposition for the patient.}

\item{earrest_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing cardiac arrest data.}

\item{evitals_06_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing systolic blood pressure (SBP) data from initial vital signs.}

\item{evitals_07_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing diastolic blood pressure (DBP) data from initial vital signs.}

\item{evitals_10_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing heart rate data from initial vital signs.}

\item{evitals_12_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing spO2 data from the initial vital signs.}

\item{evitals_14_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing respiratory rate data from initial vital signs.}

\item{evitals_23_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing total Glasgow Coma Scale (GCS) scores from initial vital signs.}

\item{evitals_26_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing alert, verbal, painful, unresponsive (AVPU) vital signs.}

\item{...}{Additional arguments passed to the \code{summarize_measure} function.}
}
\value{
A tibble summarizing results for three population groups (Adults, and Peds) with the following columns:

\code{pop}: Population type (Adults, Peds).
\code{numerator}: Count of incidents where all applicable vital signs are taken.
\code{denominator}: Total count of incidents.
\code{prop}: Proportion of incidents where all applicable vital signs are taken.
\code{prop_label}: Proportion formatted as a percentage with a specified number of
decimal places.
}
\description{
This function calculates the TTR_01 measure, which evaluates the completeness of vitals documentation for patients not experiencing cardiac arrest who were also not transported during a 911 response. It determines the total population, adult population, and pediatric population meeting the criteria for the TTR_01 measure.
}
\section{Data Assumptions}{


This function assumes the input dataset contains the \emph{initial} vital signs for
\itemize{
\item respiratory rate,
\item systolic blood pressure (SBP),
\item diastolic blood pressure (DBP),
\item heart rate (HR),
\item pulse oximetry (sp02), and
\item total Glasgow Coma Scale (GCS) score, respectively.
\item Complete AVPU responses are expected.
}
}

\author{
Nicolas Foss, Ed.D., MS
}
