% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttr_01_population.R
\name{ttr_01_population}
\alias{ttr_01_population}
\title{TTR-01 Populations}
\usage{
ttr_01_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  disposition_table = NULL,
  vitals_table = NULL,
  arrest_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  transport_disposition_col,
  earrest_01_col,
  evitals_06_col,
  evitals_07_col,
  evitals_10_col,
  evitals_12_col,
  evitals_14_col,
  evitals_23_col,
  evitals_26_col,
  ...
)
}
\arguments{
\item{df}{A data frame or tibble containing the dataset to analyze. Default is \code{NULL}.}

\item{patient_scene_table}{A data frame or tibble containing only epatient and escene fields as a fact table. Default is \code{NULL}.}

\item{response_table}{A data frame or tibble containing only the eresponse fields needed for this measure's calculations. Default is \code{NULL}.}

\item{disposition_table}{A data frame or tibble containing only the edisposition fields needed for this measure's calculations. Default is \code{NULL}.}

\item{vitals_table}{A data frame or tibble containing only the evitals fields needed for this measure's calculations. Default is \code{NULL}.}

\item{arrest_table}{A data frame or tibble containing only the earrest fields needed for this measure's calculations. Default is \code{NULL}.}

\item{erecord_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column specifying unique patient records.}

\item{incident_date_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column indicating the incident date. Must be of class \code{Date} or \code{POSIXct}. Default is \code{NULL}.}

\item{patient_DOB_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column indicating the patient's date of birth. Must be of class \code{Date} or \code{POSIXct}. Default is \code{NULL}.}

\item{epatient_15_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column indicating the patient’s age in numeric form.}

\item{epatient_16_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column specifying the unit of patient age (e.g., "Years", "Days").}

\item{eresponse_05_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column specifying the type of response (e.g., 911 codes).}

\item{transport_disposition_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column specifying transport disposition for the patient.}

\item{earrest_01_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing cardiac arrest data.}

\item{evitals_06_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing systolic blood pressure (SBP) data from initial vital signs.}

\item{evitals_07_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing diastolic blood pressure (DBP) data from initial vital signs.}

\item{evitals_10_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing heart rate data from initial vital signs.}

\item{evitals_12_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing spO2 data from the initial vital signs.}

\item{evitals_14_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing respiratory rate data from initial vital signs.}

\item{evitals_23_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing total Glasgow Coma Scale (GCS) scores from initial vital signs.}

\item{evitals_26_col}{<\link[=dplyr_tidy_select]{'tidy-select'}> A column containing alert, verbal, painful, unresponsive (AVPU) vital signs.}
}
\value{
#' A list that contains the following:
\itemize{
\item a tibble with counts for each filtering step,
\item a tibble for each population of interest
\item a tibble for the initial population
}
}
\description{
Filters data down to the target populations for TTR-01, and categorizes
records to identify needed information for the calculations.
}
\details{
Identifies key categories to records that are 911 requests for patients not transported
by EMS during which a basic set of vital signs is documented based on specific criteria
and calculates related ECG measures. This function segments the data by age into adult and pediatric populations.
}
\section{Data Assumptions}{


This function assumes the input dataset contains the \emph{initial} or complete set of vital signs for
\itemize{
\item respiratory rate,
\item systolic blood pressure (SBP),
\item diastolic blood pressure (DBP),
\item heart rate (HR),
\item pulse oximetry (sp02), and
\item total Glasgow Coma Scale (GCS) score, respectively.
\item Complete AVPU responses are expected.
}
}

\author{
Nicolas Foss, Ed.D., MS
}
